% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R
\name{amount_equipotentialintervals}
\alias{amount_equipotentialintervals}
\title{Decide on number of equipotential intervals}
\usage{
amount_equipotentialintervals(df, dp, Nf = 5)
}
\arguments{
\item{df}{tibble with information about domain}

\item{dp}{tibble with head and flow potential results for each node}

\item{Nf}{number of requested flow channels}
}
\description{
This function takes a finite difference solution for both head `h` and flow
potential function `psi` to get the best number of equipotential intervals
`Nd` based on a given number of flow channels `Nf`. This function assumes
the permeabilities in all domains is the same
}
\examples{
df <- generate_flownet_properties()
dp <- solve_flownet(df)
dp <- flow_potential(df, dp)
amount_equipotentialintervals(df, dp, Nf = 5)
}
