% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shallow_foundations.R
\name{calculate_bearingcapacityfactors}
\alias{calculate_bearingcapacityfactors}
\title{Shallow foundation bearing capacity factors}
\usage{
calculate_bearingcapacityfactors(phi)
}
\arguments{
\item{phi}{angle of internal friction, in radians (array)}
}
\value{
tibble with fields `phi` for angle of internal friction, and
  `Nc`, `Nq` and `Ngamma` for the three bearing capacity factors
}
\description{
Calculate shallow foundation bearing capacity factors, based on the
value of the friction angle, according to EC7
}
