% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidation_time.R
\name{calculate_consolidation_averagetime}
\alias{calculate_consolidation_averagetime}
\title{Calculate average consolidation in layer as function of time}
\usage{
calculate_consolidation_averagetime(Tv, trace = 1, nfourier = 50)
}
\arguments{
\item{Tv}{normalised time for crosshair annotation}

\item{trace}{trace selection (1, 2 or 3) for crosshair annotation}

\item{nfourier}{number of fourier steps to use}
}
\value{
an array with average degree of consolidation Uv
}
\description{
calculate average consolidation in a layer as function of time.
Calculations approximate the solution using a Taylor series,
similar function `calculate_consolidation_openlayer()`.

Three cases are distinguished:
* `trace = 1` drainage in open layers with thickness 2d (uniform or
    triangular), or in half-closed layer with uniform initial pressure
    and thickness d
* `trace = 2` drainage in half-closed layer with triangular distribution
    of initial excess pore water pressure. Larger pressure near closed
    end
* `trace = 3` drainage in half-closed layer with triangular distribution
    of initial excess pore water pressure. Larger pressure near open
    end
}
