% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidation_time.R
\name{calculate_consolidation_openlayer}
\alias{calculate_consolidation_openlayer}
\title{Double-drained consolidation at point in time and space}
\usage{
calculate_consolidation_openlayer(zd, Tv, nfourier = 50)
}
\arguments{
\item{zd}{normalised depth (array)}

\item{Tv}{normalised time (array)}

\item{nfourier}{number of fourier steps to use}
}
\value{
array with degrees of consolidation (between 0 and 1)
}
\description{
Calculates the degree of consolidation Uv in a double-drained
soil layer at a certain (normalised) times: Tv = c_v*t/d^2 and
at certain (normalised) position in the layer (zd = z/d), where
'd' is half the thickness of the open layer

Initial excess pore pressure is assumed to be uniform across the
entire soil layer

Solution uses the fourier series method to solve the consolidation
equation, detailed in Craig's Soil Mechanics, among others
}
