% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geometry.R
\name{circle_area_positive}
\alias{circle_area_positive}
\title{Get area of circle on positive side of a cartesian axis system}
\usage{
circle_area_positive(x, y, d)
}
\arguments{
\item{x, y}{x, y position of the circle midpoint (scalar)}

\item{d}{diameter of the circle}
}
\value{
area (scalar)
}
\description{
Calculate the area of a circle that lies on the positive side of a
two-dimensional cartesian coordinate system
}
\examples{
#complete on positive sides
circle_area_positive(1, 1, 2)

#crossing x-axis only
circle_area_positive(0, 1, 2)
circle_area_positive(-0.5, 1, 2)

#crossing y-axis only
circle_area_positive(1, 0, 2)

#crossing both axes
circle_area_positive(0.5, 0.5, 2)
circle_area_positive(-0.5, -0.5, 2)

#fully on negative side
circle_area_positive(-1, -1, 2)
}
