% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_triangle.R
\name{classify_classificationtriangle}
\alias{classify_classificationtriangle}
\title{Classify soil using BS classification triangle}
\usage{
classify_classificationtriangle(gravel, sand, silt, clay, multiple = FALSE)
}
\arguments{
\item{gravel}{mass of gravel}

\item{sand}{mass of sand}

\item{silt}{mass of silt}

\item{clay}{mass of clay}

\item{multiple}{if `TRUE`, soils that are on boundary of classifications may
belong to either side. As a result, for a single entry multiple
classifications may be returned. if `FALSE`, only one soil type is
returned, based on best interpretation of code: a) primary type only if
>40\%. Secondary type if >=15 and <=40\% for fines and >20\% and <=40\% for
sand and gravel. If equal portions, gravel takes precedence over sand,
and clay takes precendence over silt.}
}
\value{
tibble object with fractions and classification. Multiple
  classifications may be returned per input if classification lies
  on a boundary. The field `index` identifies the index of the input.
}
\description{
Add classification to soil using BS classification chart, based on measured
quantities of gravel, .data$ssand, silt and clay
}
\examples{
gravel <- c(10, 30, 50)
sand <- c(40, 20, 15)
silt <- c(30, 50, 10)
clay <- c(20, 0, 25)
classify_classificationtriangle(gravel, sand, silt, clay, multiple = FALSE)
classify_classificationtriangle(gravel, sand, silt, clay, multiple = TRUE)
}
