% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_triangle.R
\name{convert_to_grid_part1}
\alias{convert_to_grid_part1}
\title{Convert fractions to coordinates - plot 1}
\usage{
convert_to_grid_part1(gravel, sand, fines)
}
\arguments{
\item{gravel}{mass of gravel (array)}

\item{sand}{mass of sand (array)}

\item{fines}{mass of fines (array)}
}
\value{
dataframe with x and y positions
}
\description{
Function takes masses of gravel, sand and fines, and converts their
quantities to cartesian x-y positions in the triangular part of the
classification chart. x=0 and y=0 conincides with the lower left
corner, and x=1 and y=0 with the lower right corner
}
