% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_triangle.R
\name{convert_to_grid_part2}
\alias{convert_to_grid_part2}
\title{Convert fractions to coordinates - plot 2}
\usage{
convert_to_grid_part2(fines_fraction, clay_fraction)
}
\arguments{
\item{fines_fraction}{normalised mass of gravel (array)}

\item{clay_fraction}{normalised mass of sand (array)}
}
\value{
dataframe with x and y positions
}
\description{
Function takes masses of fines and clay (already normalised by total mass),
and converts their quantities to cartesian x-y positions in the
bottom square part of the classification chart. x=0 and y=0 conincides with
the lower left corner of the graph, and x=1 and y=0 with the lower right
corner.
}
