% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plotting.R
\name{depreciated_round_limits}
\alias{depreciated_round_limits}
\title{Round values in an array to a 'nice' nearby number}
\usage{
depreciated_round_limits(
  x,
  xl = c(1, 1.2, 1.5, 2, 2.5, 3, 4, 5, 6, 8, 10),
  upper = TRUE
)
}
\arguments{
\item{x}{array with values to be rounded}

\item{xl}{array with accepted values on a range [1,10].}

\item{upper}{if `upper == TRUE`, values are rounded upwards.
If `upper = FALSE`, values are rounded downwards}
}
\value{
array with rounded values. Same size as input `x`
}
\description{
Function to round an array of values to a 'nice' nearby number rounded
numbers. This can be useful for setting plot limits. Values can be
rounded upwards or downwards depending in the user choice
}
\examples{
x <- 100 * runif(5)
xf <- depreciated_round_limits(x, upper = TRUE)
print(x)
print(xf)
}
