% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R
\name{findiff_entries_bc}
\alias{findiff_entries_bc}
\title{Get finite difference matrix entries for all boundaries}
\usage{
findiff_entries_bc(df)
}
\arguments{
\item{df}{tibble with all properties for the rectangular grids.}
}
\value{
tibble with non-zero matrix entries. Fields `row` for row index,
  `col` for column index and `val` for values. These can be used to
  construct full head solver matrix later
}
\description{
Get row, column and value arrays for all non-zero entries in the problem
finite difference grid, for all boundaries of each rectangular grid section
}
