% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R
\name{findiff_entries_real}
\alias{findiff_entries_real}
\title{Get finite difference matrix entries for real nodes}
\usage{
findiff_entries_real(nx, ny, hx = 1, hy = 1, kx = 1, ky = 1, i0 = 0, ...)
}
\arguments{
\item{nx, ny}{number of nodes in x and y-direction}

\item{hx, hy}{grid size in x and y-direction}

\item{kx, ky}{permeability in x and y-direction}

\item{i0}{starting node}
}
\value{
tibble with non-zero matrix entries. Fields `row` for row index,
  `col` for column index and `val` for values. These can be used to
  construct full solver matrix later
}
\description{
Get row, column and value arrays for all non-zero entries in the second
order finite difference grid, for all real points in a grid.
}
