% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_differences.R
\name{findiff_sparse_elements_realonly}
\alias{findiff_sparse_elements_realonly}
\title{Get sparse matrix elements for 1st order differentiation (no ghost nodes)}
\usage{
findiff_sparse_elements_realonly(
  nx,
  ny,
  direction,
  i0 = 0,
  multiplier = 1,
  ...
)
}
\arguments{
\item{nx, ny}{number of real nodes in x and y-directions. There need to be
at least 3 (real or ghost) nodes in the direction of differentiation,
or 4 in case 2nd order differentiation is requested with
\code{real_only == TRUE}}

\item{direction}{direction of differentiation
\itemize{
\item \code{direction = "x"}: 1st order in x-direction (d/dx)
\item \code{direction = "y"}: 1st order in y-direction (d/dy)
}}

\item{i0}{optional index offset for node offset}

\item{multiplier}{optional multipliers for all values in the finite
difference matrix to differentiate real nodes}

\item{...}{potential extra arguments}
}
\value{
a tibble with node indices for row (\code{row}), column (\code{col}) and
value (\code{val}) columns of non-zero entries in the matrix
}
\description{
Function generates rows, column and values for sparse matrix to get a
first order approximation of values for all real nodes in a regularly
spaced, rectangular grid in either the x or y direction.

The function assumes a regular, square grid where nodes are numbered
row-first. The grid does not contain any ghost nodes. The grid has a
size 1*1.

For a 4*3 grid, the numbering is:\tabular{ccccc}{
    \tab ix=1 \tab ix=2 \tab ix=3 \tab ix=4 \cr
   iy=3 \tab 9 \tab 10 \tab 11 \tab 12 \cr
   iy=2 \tab 5 \tab 6 \tab 7 \tab 8 \cr
   iy=1 \tab 1 \tab 2 \tab 3 \tab 4 \cr
}


The finite difference approximation is of the second order (central
differences). When only real nodes are used and the node lies on an edge,
a second order approximation is used (forwards or backwards, depending
on the position)
}
\examples{
findiff_sparse_elements_realonly(9, 5, "y")
}
