% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_rectangular.R
\name{findiff_sparse_entries_bc_rectangular}
\alias{findiff_sparse_entries_bc_rectangular}
\title{get all sparse entries and lhs for boundary conditions}
\usage{
findiff_sparse_entries_bc_rectangular(df)
}
\arguments{
\item{df}{tibble describing flow problem. Generated by function
`generate_flownet_properties()`}
}
\value{
a list with two fields: `mat` containing a tibble with sparse
  row (`row`), column (`col`) and values (`val`) of finite difference
  matrix elements related to boundary conditions; and `lhs` which is a
  vector with all left-hand side of equation values
}
\description{
function to get sparse matrix element positions for flow net finite
difference boundary conditions (known head, known flow or connected to
another domain)
}
\examples{
df <- flownet_geometry_rectangular()
ds <- findiff_sparse_entries_bc_rectangular(df)
}
