% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R, R/flow_nets_new.R
\name{flow_potential}
\alias{flow_potential}
\title{Calculate the flow potential at every point}
\usage{
flow_potential(df, dp)

flow_potential(df, dp)
}
\arguments{
\item{df}{tibble with problem prescription}

\item{dp}{tibble with finite difference solution. Should contain fields for
`id` (domain identifier), `x` and `y` (for position of node) and `qx` and
`qy` for flow rate in x and y-directions}
}
\description{
Function takes the solved heads and flow from the finite difference solver
for the head, and calculates the flow potential for every real point in the
grid. This is done by solving the first-order system of equations
dpsi/dx = qy & dpsi/dy = -qx as well as extra conditions to ensure nodes on
edge of each domain match

Function takes the solved heads and flow from the finite difference solver
for the head, and calculates the flow potential for every real point in the
grid. This is done by solving the first-order system of equations
dpsi/dx = qy & dpsi/dy = -qx as well as extra conditions to ensure nodes on
edge of each domain match
}
\examples{
df <- generate_flownet_properties()
dp <- solve_flownet(df)
dp2 <- flow_potential(df, dp)
ggplot2::ggplot(
  dp2,
  ggplot2::aes(x = x, y = y, fill = psi, z = psi)) +
  ggplot2::geom_tile() +
  ggplot2::geom_contour()
df <- generate_flownet_properties()
dp <- solve_flownet(df)
dp2 <- flow_potential(df, dp)
ggplot2::ggplot(
  dp2,
  ggplot2::aes(x = x, y = y, fill = psi, z = psi)) +
  ggplot2::geom_tile() +
  ggplot2::geom_contour()
}
