% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_net.R
\name{flownet_geometry}
\alias{flownet_geometry}
\title{Define flow net problem using grided domain system}
\usage{
flownet_geometry(
  x = c(0, 15, 20, 40),
  y = c(0, 10, 12, 17.5, 20),
  ix = c(1, 1, 1, 1, 2, 3, 3),
  iy = c(4, 3, 2, 1, 1, 1, 2),
  kx = 1e-06,
  ky = 1e-06,
  bc_id = c(2, 7),
  bc_edge = c(2, 2),
  bc_type = c("h", "h"),
  bc_value = c(17.5, 15),
  dry = c(TRUE, rep(FALSE, 6)),
  grid_size = 0.25,
  node_min = 3
)
}
\arguments{
\item{x}{x-coordinates of boundary lines in the grid}

\item{y}{y-coordinates of boundary lines in the grid}

\item{ix}{x-index in grid of soil rectangle (length n)}

\item{iy}{y-index in grid of soil rectangle (length n)}

\item{kx, ky}{horizontal and vertical permeabilities. Either define as
arrays (with length n), or as scalars if the permeability is the same
in each domain}

\item{bc_id}{domain at which the boundary condition applies (array with
length m, the integers relate to the position in `ix`,`iy`)}

\item{bc_edge}{edge in domain at which the boundary condition applies (
array with length m). Edges are numbered clockwise: `1` for left
(negative x-face), `2` for top (positive y-face), `3` for right (
positive x-face) or `4` for bottom (negative y-face)}

\item{bc_type}{type of boundary condition (array with length m). Use `h`
for a known head, and `q` for a known flow (positive when flowing into
the domain)}

\item{bc_value}{Values for boundary conditions (array with length m). If
`bc_type == "h"`, it defines the known hydraulic head on the boundary.
If `bc_type == "q"`, it defines the fixed discharge into the domain on
this boundary.}

\item{dry}{if `TRUE`, the soil domain is dry (above the water table) and
will not be included in the calculations (but may be plotted). Array
size n}

\item{grid_size}{the requested grid size. Actual sizes may be slightly
larger to ensure a regular number of nodes fits in each domain. Thus
grid sizes may be slightly different in each domain}

\item{node_min}{minimum number of nodes in each direction in each
domain. Set to at least 3 to ensure finite difference approximation
works properly}
}
\description{
Defines a geometry for a flow net problem. rectangular domains are defined
using a grid system. For each domain, horizontal and vertical permeabilities
may be defined seperately.

Domains that are next to each other in the grid are automatically connected.
All boundary conditions are assumed impermeable, unless otherwise defined
using the inputs `bc_id`, `bc_edge`, `bc_type` and `bc_value`.
}
\examples{
flownet_geometry()
}
