% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_quadrilateral.R
\name{flownet_geometry_quadrilateral}
\alias{flownet_geometry_quadrilateral}
\title{Define flow net problem using gridded quadrilateral domain system}
\usage{
flownet_geometry_quadrilateral(
  x = matrix(c(1, 3, 6, 1, 3, 5, 1, 3, 5), ncol = 3, byrow = FALSE),
  y = matrix(c(0, 0, 0, 1, 1, 1.5, 2.5, 2, 2), ncol = 3, byrow = FALSE),
  ix = c(1, 1, 2),
  iy = c(2, 1, 1),
  kx = 1e-06,
  ky = 1e-06,
  grid_size = 0.05,
  bc_domain = c(1, 3),
  bc_edge = c(2, 3),
  bc_type = c("h", "h"),
  bc_value = c(10, 5),
  node_min = 3
)
}
\arguments{
\item{x}{matrix of x-coordinates of corner points in domain}

\item{y}{matrix of y-coordinates of corner points in domain}

\item{ix}{x-index in grid of soil rectangle (length n)}

\item{iy}{y-index in grid of soil rectangle (length n)}

\item{kx, ky}{horizontal and vertical permeabilities. Either define as
arrays (with length n), or as scalars if the permeability is the same
in each domain}

\item{grid_size}{the requested grid size. Actual sizes may be slightly
larger to ensure a regular number of nodes fits in each domain. Thus
grid sizes may be slightly different in each domain}

\item{bc_domain}{domain at which the boundary condition applies (array with
length m, the integers relate to the position in `ix`,`iy`)}

\item{bc_edge}{edge in domain at which the boundary condition applies (
array with length m). Edges are numbered clockwise: `1` for left
(negative x-face), `2` for top (positive y-face), `3` for right (
positive x-face) or `4` for bottom (negative y-face)}

\item{bc_type}{type of boundary condition (array with length m). Use `h`
for a known head, and `q` for a known flow (positive when flowing into
the domain)}

\item{bc_value}{Values for boundary conditions (array with length m). If
`bc_type == "h"`, it defines the known hydraulic head on the boundary.
If `bc_type == "q"`, it defines the fixed discharge into the domain on
this boundary.}

\item{node_min}{minimum number of nodes in each direction in each
domain. Set to at least 3 to ensure finite difference approximation
works properly}
}
\description{
Defines a geometry for a flow net problem. quadrilateral domains are defined
using a grid system. For each domain, horizontal and vertical permeabilities
may be defined seperately. All quadrilateral domains should be strictly
convex in shape.

Domains that are next to each other in the grid are automatically connected.
All boundary conditions are assumed impermeable, unless otherwise defined
using the inputs `bc_id`, `bc_edge`, `bc_type` and `bc_value`.
}
\examples{
flownet_geometry_quadrilateral()
}
