% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_quadrilateral.R
\name{flownet_interpolate_quadrilateral}
\alias{flownet_interpolate_quadrilateral}
\title{Interpolate quadrilateral flownet solution on rectangular grid}
\usage{
flownet_interpolate_quadrilateral(df, dp, grid_size = NULL)
}
\arguments{
\item{df}{list with geometry and boundary conditions of the problem. See
function `flownet_geometry_quadrilateral()` for more information}

\item{dp}{tibble with solution, see function `flownet_solve_quadrilateral()`}

\item{grid_size}{grid size to use for rectangular interpolation grid. If
not defined, it is chosen roughly in line with the smallest quadrilateral
grid cell}
}
\description{
Take a flow net solution calculated on a quadrilateral grid, and
interpolate the results on a regular grid. This grid can be used
to calculate contour lines (flow lines etc)
}
\examples{
df <- flownet_geometry_quadrilateral(grid_size = 0.25)
dp <- flownet_solve_quadrilateral(df)
di <- flownet_interpolate_quadrilateral(df, dp)

ggplot2::ggplot() +
  ggplot2::geom_tile(
    data = di,
    ggplot2::aes(x = x, y = y, fill = h)
  ) +
  ggplot2::geom_point(
    data = dp,
    ggplot2::aes(x = x, y = y, fill = h),
    color = "black",
    pch = 21
  ) +
  ggplot2::scale_fill_distiller(palette = "RdYlBu")
}
