% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robertsons_chart.R
\name{FrH}
\alias{FrH}
\title{Qtn-Fr relationship for boundary between zone 8 and 9}
\usage{
FrH()
}
\value{
Fr values
}
\description{
Calculate Fr based on known Qtn for line seperating zones 8 and 9
}
