% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robertsons_chart.R
\name{FrQtn2Ic}
\alias{FrQtn2Ic}
\title{Calculate Ic from known Fr and Qtn, in zone 2 - 7}
\usage{
FrQtn2Ic(Fr, Qtn)
}
\arguments{
\item{Fr}{normalised friction ratio, in \%}

\item{Qtn}{normalised tip resistance}
}
\value{
Robertson's Ic index
}
\description{
Calculate Ic from known Fr and Qtn, in zone 2 - 7 in Robertson's
chart
}
