% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plotting.R
\name{get_log10_minorbreaks}
\alias{get_log10_minorbreaks}
\title{Get all minor log tick positions in log10 axis}
\usage{
get_log10_minorbreaks(x)
}
\arguments{
\item{x}{array with minimum and maximum axis limits}
}
\description{
Function returns a list of minor axis ticks for a certain range of values
}
