% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_weights.R
\name{get_unitweight}
\alias{get_unitweight}
\title{Function to obtain unit weight from soil description}
\usage{
get_unitweight(description, round = 0.5)
}
\arguments{
\item{description}{array with descriptions}

\item{round}{round unit weights to be nearest multiple of this number}
}
\description{
Get soil unit weight, above and below the water table, from a soil
description
}
\examples{
description <- c("very dense sand", "loose gravel")
get_unitweight(description)
}
