% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_plotting.R
\name{ggplot_add_flownet}
\alias{ggplot_add_flownet}
\title{Add flownet to an existing ggplot}
\usage{
ggplot_add_flownet(
  df,
  dp,
  di = NULL,
  Nf = 5,
  linewidth = 0.5,
  colour_flowline = "black",
  colour_equipotentialline = "black",
  fill = "#aebab7",
  plt = NULL,
  xlab = "x [m]",
  ylab = "z [m]",
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  axes = TRUE
)
}
\arguments{
\item{df}{tibble with domains/problem description}

\item{dp}{flow net solution tibble}

\item{di}{flow net solution rectangular interpolation tibble. This tibble
is used for plotting contour lines (flow lines and equipotential lines).
If the grid is rectangular, this is simply equal to `dp`. If not, an
interpolation function is used. If not defined, it is calculated from
`dp`}

\item{Nf}{number of requested flow paths}

\item{linewidth}{line thickness for all lines}

\item{colour_flowline}{colour of flow lines}

\item{colour_equipotentialline}{colour of equipotential lines}

\item{fill}{fill colour for (saturated) soil}

\item{plt}{a ggplot object. If not defined, a new plot is generated}

\item{xlab, ylab}{x and y-axis label - if `plt` not defined}

\item{xlim, ylim}{user-defined axis limits - if `plt` not defined}

\item{axes}{if `FALSE` no axes are plotted - if `plt` not defined}
}
\value{
a ggplot object
}
\description{
Function to add a flow net to an existing ggplot. If the plot does not
exist yet, a new one if created
}
\examples{
#rectangular grid
df <- flownet_geometry_rectangular()
dp <- flownet_solve_rectangular(df)
ggplot_add_flownet(df, dp, axes = FALSE)
ggplot_add_flownet(df, dp, axes = TRUE, fill = NA)

#quadrilateral grid
df <- flownet_geometry_quadrilateral()
dp <- flownet_solve_quadrilateral(df)
di <- flownet_interpolate_quadrilateral(df, dp)
ggplot_add_flownet(df, dp, di = di, axes = FALSE)
}
