% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_quadrilateral.R
\name{ggplot_add_flownet_quadrilateral}
\alias{ggplot_add_flownet_quadrilateral}
\title{Add flownet (quadrilateral domains) to an existing ggplot}
\usage{
ggplot_add_flownet_quadrilateral(
  df,
  dp = NULL,
  di = NULL,
  Nf = 5,
  linewidth = 0.5,
  colour_flowline = "black",
  colour_equipotentialline = "black",
  fill = "#aebab7",
  plt = NULL,
  xlab = "x [m]",
  ylab = "z [m]",
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  axes = TRUE
)
}
\arguments{
\item{df}{tibble with domains/problem description}

\item{dp}{flow net solution tibble. If not specified, it is calculated
using `df`}

\item{di}{flow net solution interpolated on a regular rectangular
domain. If not specified, it is calculated using `dp`}

\item{Nf}{number of requested flow paths}

\item{linewidth}{line thickness for all lines}

\item{colour_flowline}{colour of flow lines}

\item{colour_equipotentialline}{colour of equipotential lines}

\item{fill}{fill colour for (saturated) soil}

\item{plt}{a ggplot object. If not defined, a new plot is generated}

\item{xlab, ylab}{x and y-axis label - if `plt` not defined}

\item{xlim, ylim}{user-defined axis limits - if `plt` not defined}

\item{axes}{if `FALSE` no axes are plotted - if `plt` not defined}
}
\value{
a ggplot object
}
\description{
Function to add a flow net to an existing ggplot. If the plot does not
exist yet, a new one if created
}
\examples{
#standard example
df <- flownet_geometry_quadrilateral(grid_size = 0.05)
dp <- flownet_solve_quadrilateral(df)
di <- flownet_interpolate_quadrilateral(df, dp)
ggplot_add_flownet_quadrilateral(df, dp, di, Nf = 6)
}
