% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plotgeometry.R
\name{ggplot_add_polygon}
\alias{ggplot_add_polygon}
\title{Add a series of polygons to a ggplot object}
\usage{
ggplot_add_polygon(
  plt,
  df,
  type = "soil",
  fill = NULL,
  settings = tibble::tibble(id = c("soil", "saturated", "water", "structure"), fill =
    c("#d3bc5f", "#aebab7", "#2a7fff", "#333333"), colour = c(NA, NA, NA, NA), )
)
}
\arguments{
\item{plt}{ggplot to be added to}

\item{df}{tibble with data. Should contain fields `x` and `y` for
polygon point positions. May contain optional field `group` to
differentiate between different polygons}

\item{type}{either `soil`, `water` or `structure`. This setting is
used to determine the fill colour of the polygons}

\item{fill}{if specified, overrides the fill colour defined through the
`settings` settings}

\item{settings}{dataframe with default settings for soil, water and
structure elements}
}
\value{
ggplot object
}
\description{
Function to make it faster to add a series of soil, water or structure
polygons to an existing ggplot object
}
