% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_crosshairs.R
\name{ggplot_addcrosshairs}
\alias{ggplot_addcrosshairs}
\title{Add datapoints/crosshairs to an existing plot}
\usage{
ggplot_addcrosshairs(
  plt,
  x,
  y,
  group = NULL,
  legend_title = NULL,
  legend_position = "none",
  crosshairs = TRUE,
  add_colour_scale = FALSE,
  xlim = 0,
  ylim = 0,
  arrow_x = TRUE,
  arrow_y = TRUE,
  arrow_length = 0.1,
  palette = "Set1",
  label_size = 4,
  label_hjust = 0,
  label_vjust = 0,
  label_nudge_x = 0,
  label_nudge_y = 0,
  label_parse = FALSE
)
}
\arguments{
\item{plt}{a ggplot object}

\item{x}{x-coordinates for all points}

\item{y}{y-coordinates for all points}

\item{group}{names of group, for all points - this is plotted in the label}

\item{legend_title}{title for legend}

\item{legend_position}{legend position. If `NULL`, no legend is drawn}

\item{crosshairs}{if `TRUE`, plot crosshairs for each data point}

\item{add_colour_scale}{if `TRUE`, add new color scale for different crosshairs}

\item{xlim}{start point of x crosshairs}

\item{ylim}{start point of y crosshairs}

\item{arrow_x, arrow_y}{if `TRUE`, point x or y-arrows towards the point.
If `FALSE`, points towards the axis}

\item{arrow_length}{length of arrow heads}

\item{palette}{RColorBrewer color palette to use}

\item{label_size}{text size in label}

\item{label_hjust}{horizontal justification of labels}

\item{label_vjust}{vertical justification of labels}

\item{label_nudge_x}{x-offset of labels}

\item{label_nudge_y}{y-offset of labels}

\item{label_parse}{if `TRUE`, parse label text}
}
\description{
Function takes an existing ggplot and add points and/or crosshairs to the
plot to show the position of points
}
