% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shallow_foundations.R
\name{ggplot_bearingcapacityfactors}
\alias{ggplot_bearingcapacityfactors}
\title{ggplot chart with foundation bearing capacity factors}
\usage{
ggplot_bearingcapacityfactors(
  phi_deg = seq(0, 50, 1),
  ylim = c(1, 1000),
  palette = "Set1"
)
}
\arguments{
\item{phi_deg}{array with all values for the angle of internal friction,
in degrees (array)}

\item{ylim}{2-value array with y-axis limits}

\item{palette}{RCOlorBrewer color palette to use for line colors}
}
\value{
ggplot object
}
\description{
Creates an ggplot chart with the shallow foundation bearing
capacity factors, varying with the angle of internal friction, according
to EC7
}
\examples{
ggplot_bearingcapacityfactors()
}
