% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oedometer.R
\name{ggplot_casagrande_preconsolidation}
\alias{ggplot_casagrande_preconsolidation}
\title{ggplot for Casagrande oedometer preconsolidation pressure}
\usage{
ggplot_casagrande_preconsolidation(
  sigma_v = c(5, 10, 20, 50, 100, 200, 500, 1000, 2000),
  e = c(0.94, 0.93, 0.92, 0.9, 0.88, 0.85, 0.8, 0.75, 0.71),
  xlim = c(4, NA),
  ylim = c(NA, NA),
  palette = "Set1",
  label_line = c("1", "2", "3", "4", "5", "sigma*minute[v*\\",\\"*c]"),
  stages = seq(6),
  include_value = TRUE,
  nsignif = 2
)
}
\arguments{
\item{sigma_v}{array with oedometer stresses}

\item{e}{array with void ratios}

\item{xlim, ylim}{2-parameters arrays for x and y axes limits. If not
defined, these limits are determined automatically}

\item{palette}{RColorBrewer pallette for plot colors}

\item{label_line}{array with six character strings, which will be parsed
\itemize{
\item string for tangent line at maximum curvature (1)
\item string for horizontal line (2)
\item string for bisection line between lines (1) and (2)
\item string for tangent line virgin behaviour (3)
\item string for vertical line down from intersection lines (3) and (4)
\item string for preconsolidation pressure label
}}

\item{stages}{plot fitting stages (array with numbers)
\itemize{
\item \code{1}: adds smooth trace through measurement points
\item \code{2}: adds trangent at max curvature
\item \code{3}: adds horizontal line at max curvature
\item \code{4}: adds bisection line between 2 and 3
\item \code{5}: adds tangent at virgin compression
\item \code{6}: adds vertical at intersection of 4 and 5
}}

\item{include_value}{if \code{TRUE}, the value of the preconsolidation pressure
is added to the plotted label}

\item{nsignif}{the number of significant digits to use in the (rounded)
value of the show preconsolidation pressure (if \code{include_value = TRUE})}
}
\description{
Function to create a ggplot showing how Casagrande's method for determining
the preconsolidation pressure from oedometer data works.

The point of maximum curvature is determined by fitting a bilinear curve
with a hyperbolic transition area.
}
\examples{
#all stages
ggplot_casagrande_preconsolidation()

#stage by stage
ggplot_casagrande_preconsolidation(stages = c())
ggplot_casagrande_preconsolidation(stages = seq(1))
ggplot_casagrande_preconsolidation(stages = seq(2))
ggplot_casagrande_preconsolidation(stages = seq(3))
ggplot_casagrande_preconsolidation(stages = seq(4))
ggplot_casagrande_preconsolidation(stages = seq(5))
ggplot_casagrande_preconsolidation(stages = seq(6))
}
