% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_triangle.R
\name{ggplot_classificationtriangle_addcrosshairs_part2}
\alias{ggplot_classificationtriangle_addcrosshairs_part2}
\title{Add data to classification chart - part 2}
\usage{
ggplot_classificationtriangle_addcrosshairs_part2(
  plt,
  fines_fraction,
  clay_fraction,
  group = NULL,
  crosshairs = TRUE,
  legend_position = "none",
  legend_title = "",
  palette = "Set1",
  label_parse = FALSE
)
}
\arguments{
\item{plt}{a ggplot object with chart}

\item{fines_fraction}{mass fraction of fines (array)}

\item{clay_fraction}{mass fraction of clay (array)}

\item{group}{name of group}

\item{crosshairs}{if `TRUE`, crosshairs are plotted}

\item{legend_position}{position of legend. If `none`, no legend is shown}

\item{legend_title}{title of legend}

\item{palette}{RColorBrewer palette used for discrete colors}

\item{label_parse}{if `TRUE`, parse `group` labels}
}
\value{
ggplot object
}
\description{
Add measured data to part 2 of the soil classification chart
}
