% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_triangle.R
\name{ggplot_classificationtriangle_part1}
\alias{ggplot_classificationtriangle_part1}
\title{Create soil classification chart - part 1}
\usage{
ggplot_classificationtriangle_part1(
  gravel = NULL,
  sand = NULL,
  fines = NULL,
  group = NULL,
  grid_ticks = 10,
  grid_label = 20,
  tick_length = 2,
  label_length = 7,
  label_size = 2,
  title_length = 15,
  title_size = 3,
  soil_size = 2,
  soil_size2 = 1.4,
  soil_spacing = 0.75,
  polygon_line = 0.4,
  polygon_color = "grey60",
  xlim = c(-0.2, 1.2),
  ylim = c(-0.2, 1)
)
}
\arguments{
\item{gravel}{fraction of gravel for crosshair annotation}

\item{sand}{fraction of sand for crosshair annotation}

\item{fines}{fraction of fines for crosshair annotation}

\item{group}{grouping for crosshair annotation (label)}

\item{grid_ticks}{distance between axis ticks, in \%}

\item{grid_label}{distance between axis tick labels, in \%}

\item{tick_length}{length of tick mark, in \%}

\item{label_length}{distance of tick label to chart axis, in \%}

\item{label_size}{font size of tick label}

\item{title_length}{distance to axis title to chart, in \%}

\item{title_size}{font size of axis title}

\item{soil_size}{font size of soil labels}

\item{soil_size2}{font size of soil labels, for the small ones in the
middle of the triangular plot}

\item{soil_spacing}{line spacing for soil labels}

\item{polygon_line}{thickness of soil polygon lines}

\item{polygon_color}{color of soil polygon lines}

\item{xlim}{plot x-limits}

\item{ylim}{plot y-limits}
}
\value{
a ggplot object
}
\description{
Function returns ggplot of the first parts of the BS soil classification
chart: the triagular upper part
}
\examples{
#empty chart
ggplot_classificationtriangle_part1()

#chart with some annotations
ggplot_classificationtriangle_part1(
  gravel = 0.25,
  sand = 0.30,
  fines = 0.45
)
}
