% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidation_time.R
\name{ggplot_consolidation_averageprofiles}
\alias{ggplot_consolidation_averageprofiles}
\title{ggplot profile types for average layer consolidation times}
\usage{
ggplot_consolidation_averageprofiles(
  xmarg = 0.2,
  ymarg = 0.15,
  ymarg2 = 0.3,
  fill_pressure = "#2a7fff",
  color_layer = c("white", "black"),
  fill_layer = c("grey20", "#d3bc5f"),
  label_size = 3.5,
  case_names = c("Case A", "Case B", "Case C"),
  drain_names = c("Permeable", "Impermeable"),
  layer_names = c("Open layer", "Half-closed layer"),
  arrow_size = 0.4
)
}
\arguments{
\item{xmarg}{relative horizontal offset of pressure profiles}

\item{ymarg}{relative thickness of soil layers on bottom and top}

\item{ymarg2}{relative distance between top and pressure arrow}

\item{fill_pressure}{fill color of pore water pressure profiles}

\item{color_layer}{text color of soil layer type annotation
(impermeable, permeable)}

\item{fill_layer}{fill color of soil layer annotations
(impermeable, permeable)}

\item{label_size}{text font size for all labels}

\item{case_names}{names for cases: A, B and C}

\item{drain_names}{names for drainage types: impermeable and permeable}

\item{layer_names}{layer types: half-closed layer and open layer}

\item{arrow_size}{length of arrow heads, in unit 'lines'}
}
\value{
a ggplot object
}
\description{
Creates a ggplot object with the various profiles of initial excess
pore water pressure (uniform, triangular) in open and half-closed
soil layers. The average consolidation with time can be plotted with
the function `ggplot_consolidation_averagetime()`
}
\examples{
ggplot_consolidation_averageprofiles()
}
