% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidation_time.R
\name{ggplot_consolidation_doubledrained}
\alias{ggplot_consolidation_doubledrained}
\title{ggplot consolidation in double-drained layer in time and space}
\usage{
ggplot_consolidation_doubledrained(
  zd = NULL,
  Tv = NULL,
  zd_calc = seq(0, 2, l = 101),
  Tv_calc = c(seq(0.05, 0.2, 0.05), seq(0.3, 0.9, 0.1)),
  label_zd = 0.7,
  label_size = 3,
  label_alpha = 1,
  nround = 2
)
}
\arguments{
\item{zd}{normalised depth for crosshair annotations}

\item{Tv}{normalised time for crosshair annotation}

\item{zd_calc}{normalised depth used to plot lines}

\item{Tv_calc}{normalised time steps for which to plot lines}

\item{label_zd}{depth position at which to plot time labels on traces}

\item{label_size}{size of the text in time labels}

\item{label_alpha}{opacity of the time labels}

\item{nround}{number of decimals to use in rounding values in crosshair
annotations}
}
\value{
a ggplot object
}
\description{
Plots a ggplot with development of consolidation in a double-drained
layer, in space and time.
}
\examples{
#empty chart
ggplot_consolidation_doubledrained()

#annotated chart
ggplot_consolidation_doubledrained(zd = 0.6, Tv = 0.3)
}
