% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oedometer.R
\name{ggplot_cv_sqrt}
\alias{ggplot_cv_sqrt}
\title{ggplot cv determination sqrt method}
\usage{
ggplot_cv_sqrt(
  t = c(seq(10, 60, 10), c(1, 2, 4, 8, 15, 30) * 60, c(1, 2, 4, 8, 24) * 60 * 60),
  h0 = 0.0195 * 1000,
  dsigmav = 20000 * 1e-06,
  mv = 5e-06 * 1e+06,
  cv = 1e-08 * 1e+06,
  Calpha = 0.005,
  Tvalpha = 10^0.1,
  xlim = c(NA, 150),
  ylim = c(NA, NA),
  palette = "Set1",
  line = c(2, 4, 5),
  shape = c(15, 15, 15)
)
}
\arguments{
\item{t}{time points for individual measurement points}

\item{h0}{initial sample height [m]}

\item{dsigmav}{increase in vertical stress [Pa]}

\item{mv}{coefficient of compressibility [m^2/N]}

\item{cv}{coefficient of consolidation [m^2/s]}

\item{Calpha}{creep strain per log10-cycle of time. Creep is only added
after primary consolidation is (almost) finished. A small smoothing
is applied to smoothen the sudden jump in gradient of settlement}

\item{Tvalpha}{value of normalised strain at which creep stats to be
included}

\item{xlim}{manual x and y-axis limits}

\item{ylim}{manual x and y-axis limits}

\item{palette}{RColorBrewer palette to use for colours of annotations}

\item{line}{linestyles for lines in the three steps}

\item{shape}{marker shape for the three steps}
}
\value{
ggplot object
}
\description{
ggplot showing how to obtain t50 from oedometer data using the sqrt
method
}
\examples{
#default example
ggplot_cv_sqrt()
}
