% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasticity.R
\name{ggplot_fallcone}
\alias{ggplot_fallcone}
\title{ggplotly plasticity fall cone chart: water content - penetration depth}
\usage{
ggplot_fallcone(
  w = c(9, 13, 16, 20),
  u = c(12, 17, 21, 25),
  nround = 0,
  palette = "Set1"
)
}
\arguments{
\item{w}{measured water contents, in \%}

\item{u}{measured penetration depths, in mm}

\item{nround}{number of decimals in rounded liquid limit}

\item{palette}{RColorBrewer color palette to use for colors}
}
\value{
ggplot object
}
\description{
Plots a chart with measured water content versus fall cone penetration
depth in a fall cone test. The best linear fit is plotted and the
liquid limit (rounded) is indicated
}
\examples{
w = c(9, 13, 16, 20)
u = c(12, 17, 21, 25)
ggplot_fallcone(w, u, nround = 0)
}
