% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_net.R
\name{ggplot_flownet_dam}
\alias{ggplot_flownet_dam}
\title{Plot flow net and soil geometry for a dam/sheet pile problem}
\usage{
ggplot_flownet_dam(
  df,
  dp = NULL,
  Nf = 5,
  xlab = "x [m]",
  ylab = "z [m]",
  fill_soilwet = "#aebab7",
  fill_soildry = "#d3bc5f",
  colour_soil = "#65571d",
  colour_flowline = "black",
  colour_equipotentialline = "black",
  fill_water = "#2a7fff",
  colour_water = "#000080",
  fill_impermeable = "#333333",
  fill_air = "white",
  line_water = 2,
  line_width = 0.5,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  axes = TRUE,
  thickness_impermeable = c(0, 1, 0, 1),
  scale_soilmarker = 0.075,
  scale_watermarker = 0.05
)
}
\arguments{
\item{df}{tibble with domains/problem description, generated by function
`generate_flownet_properties()`.#'}

\item{dp}{solution for heads and flow potential for each node. If not
defined, it is calculated in this plotting function}

\item{Nf}{number of requested flow paths}

\item{xlab, ylab}{x and y-axis label}

\item{fill_soilwet}{fill colour for saturated soil}

\item{fill_soildry}{fill colour for dry soil,}

\item{colour_soil}{line colour for soil}

\item{colour_flowline}{colour of flow lines}

\item{colour_equipotentialline}{colour of equipotential lines}

\item{fill_water}{fill colour for water}

\item{colour_water}{line colour for water}

\item{fill_impermeable}{fill colour for impermeable parts of the domain}

\item{fill_air}{fill colour for air on top of soil/water}

\item{line_water}{line type for water edges (defined head)}

\item{line_width}{line thickness for all lines}

\item{xlim, ylim}{user-defined axis limits}

\item{axes}{if `FALSE` no axes are plotted}

\item{thickness_impermeable}{thickness of impermeable border around the
entire problem. Defined as an array for all edges (left, top, right,
bottom), or single value for all edges}

\item{scale_soilmarker}{size of soil marker, in fraction of height}

\item{scale_watermarker}{size of water table marker, in fraction of height}
}
\value{
a ggplot object
}
\description{
Plot flow net and soil geometry for a dam/sheet pile problem
}
\examples{
#default example
df <- flownet_geometry()
ggplot_flownet_dam(df)

#anisotropic permeability - with 6 flow channels
df <- flownet_geometry(kx = 4e-6, ky = 1e-6)
ggplot_flownet_dam(df, Nf = 6)

#deep excavation example
df <- flownet_geometry(
  x = c(0, 5, 5.25, 30),
  y = c(0, 5, 7.5, 15, 20),
  ix = c(1, 1, 2, 3, 3, 3, 3),
  iy = c(2, 1, 1, 1, 2, 3, 4),
  bc_id = c(1, 6),
  bc_edge = c(2, 2),
  bc_type = c("h", "h"),
  bc_value = c(7.5, 15),
  dry = c(rep(FALSE, 6), TRUE)
)
ggplot_flownet_dam(df)
}
