% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R
\name{ggplot_flownet_only}
\alias{ggplot_flownet_only}
\title{Plot flow net in part of soil that is saturated}
\usage{
ggplot_flownet_only(
  df,
  dp = NULL,
  Nf = 5,
  xlab = "x [m]",
  ylab = "z [m]",
  fill_soilwet = "#aebab7",
  fill_soildry = "#d3bc5f",
  colour_soil = "#65571d",
  colour_flowline = "black",
  colour_equipotentialline = "black",
  fill_water = "#2a7fff",
  colour_water = "#000080",
  line_water = 2,
  line_width = 0.5,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  axes = TRUE
)
}
\arguments{
\item{df}{tibble with domains/problem description}

\item{dp}{flow net solution tibble. If not specified, it is calculated
using `df`}

\item{Nf}{number of requested flow paths}

\item{xlab, ylab}{x and y-axis label}

\item{fill_soilwet}{fill colour for saturated soil}

\item{fill_soildry}{fill colour for dry soil,}

\item{colour_soil}{line colour for soil}

\item{colour_flowline}{colour of flow lines}

\item{colour_equipotentialline}{colour of equipotential lines}

\item{fill_water}{fill colour for water}

\item{colour_water}{line colour for water}

\item{line_water}{line type for water edges (defined head)}

\item{line_width}{line thickness for all lines}

\item{xlim, ylim}{user-defined axis limits}

\item{axes}{if `FALSE` no axes are plotted}
}
\value{
a ggplot object
}
\description{
Plot flow net in part of soil that is saturated
}
\examples{
df <- generate_flownet_properties()
dp <- solve_flownet(df)
dp <- flow_potential(df, dp)
ggplot_flownet_only(df, dp, axes = FALSE)
}
