% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plotgeometry.R
\name{ggplot_geometry}
\alias{ggplot_geometry}
\title{ggplot a soil + water + structure geometry}
\usage{
ggplot_geometry(
  pol_water = NULL,
  pol_soil = NULL,
  pol_saturated = NULL,
  pol_structure = NULL,
  surf_soil = NULL,
  surf_water = NULL,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  axes = TRUE,
  xlab = "x [m]",
  ylab = "y [m]"
)
}
\arguments{
\item{pol_water}{tibble with surface water polygons}

\item{pol_soil}{tibble with soil polygons}

\item{pol_saturated}{tibble with saturated soil polygons}

\item{pol_structure}{tibble with structure polygons}

\item{surf_soil}{tibble with soil surfaces}

\item{surf_water}{tibble with water tables}

\item{xlim}{x-axis limits (array with min and max)}

\item{ylim}{y-axis limits (array with min and max)}

\item{axes}{if `TRUE`, plot axis ticks and labels. If `FALSE`, no axis
are plotted}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
ggplot object
}
\description{
Wrapper function to plot a soil, water and structure geometry
}
\examples{
#input
pol_soil = data.frame(x = c(0, 0, 2, 2), y = c(0, 1, 1.1, 0), group = 0)
pol_saturated = data.frame(x = c(0, 0, 1, 1), y = c(0, 0.5, 0.5, 0))
pol_water = data.frame(x = c(0, 0, 2, 2), y = c(1, 1.5, 1.2, 1))
pol_structure = data.frame(x = c(0.5, 1, 1), y = c(0.5, 0.5, 0.75))
surf_soil = data.frame(x = c(0, 2), y = c(1, 1.1))
surf_water = data.frame(x = c(0, 2), y = c(1.5, 1.2))

#plot
ggplot_geometry(
  pol_water = pol_water,
  pol_soil = pol_soil,
  pol_saturated = pol_saturated,
  pol_structure = pol_structure,
  surf_soil = surf_soil,
  surf_water = surf_water
)
}
