% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_net.R
\name{ggplot_hydraulichead_dam}
\alias{ggplot_hydraulichead_dam}
\title{ggplot 2D map of hydraulic heads and pressures}
\usage{
ggplot_hydraulichead_dam(
  df,
  dp = NULL,
  type = "h",
  binwidth = 0.25,
  gamma_w = 10,
  label_size = 3,
  label_n = 1,
  xlab = "x [m]",
  ylab = "z [m]",
  palette = "Blues",
  palette_direction = 1,
  fill_soilwet = "white",
  fill_soildry = "white",
  colour_soil = "#65571d",
  colour_contour = "black",
  fill_water = "white",
  colour_water = "#000080",
  fill_impermeable = "#333333",
  fill_air = "white",
  line_water = 2,
  line_width = 0.5,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  axes = TRUE,
  thickness_impermeable = c(0, 1, 0, 1),
  scale_soilmarker = 0.075,
  scale_watermarker = 0.05
)
}
\arguments{
\item{df}{tibble with domains/problem description, generated by function
`generate_flownet_properties()`.#'}

\item{dp}{solution for heads and flow potential for each node. If not
defined, it is calculated in this plotting function}

\item{type}{parameter to plot. `type = "h"` for hydraulic heads,
`type = "hz"` for elevation head, `type = `hb` for pressure head
or `type = "u"` for pore water pressures}

\item{binwidth}{contour line distance}

\item{gamma_w}{unit weight of water, in kN/m3 (required for calculating
pore water pressures)}

\item{label_size}{contour line text label size}

\item{label_n}{number of times contour label appears on each line}

\item{xlab}{x and y-axis label}

\item{ylab}{x and y-axis label}

\item{palette}{RColorBrewer color map for shading}

\item{palette_direction}{RColorBrewer color map direction (`1` or `-1`)}

\item{fill_soilwet}{fill colour for saturated soil}

\item{fill_soildry}{fill colour for dry soil,}

\item{colour_soil}{line colour for soil}

\item{colour_contour}{colour of contour lines}

\item{fill_water}{fill colour for water}

\item{colour_water}{line colour for water}

\item{fill_impermeable}{fill colour for impermeable parts of the domain}

\item{fill_air}{fill colour for air on top of soil/water}

\item{line_water}{line type for water edges (defined head)}

\item{line_width}{line thickness for all lines}

\item{xlim}{user-defined axis limits}

\item{ylim}{user-defined axis limits}

\item{axes}{if `FALSE` no axes are plotted}

\item{thickness_impermeable}{thickness of impermeable border around the
entire problem. Defined as an array for all edges (left, top, right,
bottom), or single value for all edges}

\item{scale_soilmarker}{size of soil marker, in fraction of height}

\item{scale_watermarker}{size of water table marker, in fraction of height}
}
\value{
a ggplot object
}
\description{
Plots a 2D map of the pressure or head distributions in a flow net problem.
The plot shows both a heat map and contour lines.
}
\examples{
#generate flow net geometry
df <- flownet_geometry()

#solve flow net problem
dp <- solve_flownet(df)

#plot heads
ggplot_hydraulichead_dam(df, dp = dp, type = "h", binwidth = 0.25)

#plot elevation head
ggplot_hydraulichead_dam(df, dp = dp, type = "hz", binwidth = 2.5)

#plot pressure head
ggplot_hydraulichead_dam(df, dp = dp, type = "hb", binwidth = 2.5)

#plot pore water pressure
ggplot_hydraulichead_dam(df, dp = dp, type = "u", binwidth = 25)
}
