% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_foundations.R
\name{ggplot_hyperbolicmethod}
\alias{ggplot_hyperbolicmethod}
\title{ggplot exampe pile settlement using the hyperbolic method}
\usage{
ggplot_hyperbolicmethod(
  Qb = 1500,
  Qs = 1000,
  Q = 2000,
  Eb = 50000,
  Ds = 0.5,
  Db = NULL,
  Ms = 0.001,
  ylim = c(0, 0.5),
  palette = "Set1"
)
}
\arguments{
\item{Qb}{pile base capacity}

\item{Qs}{pile shaft capacity}

\item{Q}{Example load applied to pile}

\item{Eb}{soil Young's modulus underneath the pile tip}

\item{Ds}{pile shaft diameter}

\item{Db}{pile base diameter. If not defined, this is assumed to be equal
to the shaft diameter}

\item{Ms}{empirical factor, assumed as Ms = 0.001 by default}

\item{ylim}{vector with upper and lower plot settlement limits}

\item{palette}{RColorBrewer color palette to use for annotations}
}
\description{
Returns an example pile load-settlement graph for a rigid pile using
Fleming's (1992) hyperbolic method for pile displacements.
}
\examples{
# Input parameters
Qs <- 1000  # [kN]
Qb <- 1500  # [kN]
Q <- 2000  # [kN]
Eb <- 50e3  # [kPa]
Ds <- 0.5  # [m]

# plot
ggplot_hyperbolicmethod(Qb = Qb, Qs = Qs, Q = Q, Eb = Eb, Ds = Ds)
}
