% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_foundations.R
\name{ggplot_hyperbolicmethod_stiffness}
\alias{ggplot_hyperbolicmethod_stiffness}
\title{ggplot Poulos's chart for friction along pile in sand}
\usage{
ggplot_hyperbolicmethod_stiffness()
}
\value{
ggplot object
}
\description{
Create a ggplot chart for the stiffness under the pile tip as used in the
Hyperbolic method by Fleming (1992). Data digitised from Craig's Soil
Mechanics

This function uses data from a digitised plot, so may not be 100% accurate
}
\examples{
#plot chart
ggplot_hyperbolicmethod_stiffness()
}
