% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc_plot.R
\name{ggplot_mcc_eps1_q}
\alias{ggplot_mcc_eps1_q}
\title{ggplot MCC axial strain versus deviatoric stress}
\usage{
ggplot_mcc_eps1_q(df, palette = "Set1", e1_max = NULL)
}
\arguments{
\item{df}{tibble with all MCC predictions}

\item{palette}{RColorBrewer palette for annotation colours}

\item{e1_max}{maximum strain to plot}
}
\value{
a ggplot object
}
\description{
Plot axial strain versus deviatoric stress
}
\examples{
#drained
df <- mcc_solve_triax_drained()
ggplot_mcc_eps1_q(df)
ggplot_mcc_eps1_q(df, e1_max = 0.04)

#undrained
df <- mcc_solve_triax_undrained()
ggplot_mcc_eps1_q(df)
}
