% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc_plot.R
\name{ggplot_mcc_lnp_v}
\alias{ggplot_mcc_lnp_v}
\title{ggplot MCC ln(p)-v stress paths}
\usage{
ggplot_mcc_lnp_v(
  df,
  Gamma = 2,
  lambda = 0.115,
  kappa = 0.015,
  palette = "Set1",
  e1_max = NULL
)
}
\arguments{
\item{df}{tibble with all MCC predictions}

\item{Gamma}{MCC CSL intercept specific volume}

\item{lambda, kappa}{MCC compression parameters}

\item{palette}{RColorBrewer palette for annotation colours}

\item{e1_max}{maximum strain to plot}
}
\value{
a ggplot object
}
\description{
Plot (effective) ln(p)-v stress paths
}
\examples{
Gamma <- 2
lambda <- 0.115
kappa <- 0.015
df <- mcc_solve_triax_drained(Gamma = Gamma, lambda = lambda, kappa = kappa)
ggplot_mcc_lnp_v(df, Gamma = Gamma, lambda = lambda, kappa = kappa)
ggplot_mcc_lnp_v(df, Gamma = Gamma, lambda = lambda, kappa = kappa, e1_max = 0.01)
}
