% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc_plot.R
\name{ggplot_mcc_p_q}
\alias{ggplot_mcc_p_q}
\title{ggplot MCC p-q stress paths}
\usage{
ggplot_mcc_p_q(df, M = 1.35, pc = 100, palette = "Set1", e1_max = NULL)
}
\arguments{
\item{df}{tibble with all MCC predictions}

\item{M}{MCC M-parameter}

\item{pc}{initial preconsolidation pressure}

\item{palette}{RColorBrewer palette for annotation colours}

\item{e1_max}{maximum strain to plot}
}
\value{
a ggplot object
}
\description{
Plot (effective) p-q stress paths
}
\examples{
M <- 1.35
pc <- 100

#drained
df <- mcc_solve_triax_drained(M = M, pc = pc)
ggplot_mcc_p_q(df, M = M, pc = pc)
ggplot_mcc_p_q(df, e1_max = 0.01)

#undrained
df <- mcc_solve_triax_undrained(M = M, pc = pc, p0 = 80, nu = 0.3)
ggplot_mcc_p_q(df, M = M, pc = pc)
}
