% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohr_circles.R
\name{ggplot_mohrcircle}
\alias{ggplot_mohrcircle}
\title{ggplot to plot Mohr circle of rotated stress element}
\usage{
ggplot_mohrcircle(
  sigz = 40,
  sigx = 20,
  tau = 5,
  theta = 0,
  rotation_label = "theta",
  face_label = c("X", "Z"),
  palette = "Set1",
  pole_label = "Pole",
  n_circle = 181,
  color_circle = "black",
  color_lines = "grey50",
  effective_stress = TRUE
)
}
\arguments{
\item{sigz}{normal stress on z-plane (scalar)}

\item{sigx}{normal stress on x-plane (scalar)}

\item{tau}{shear stress}

\item{theta}{rotation, in radians (scalar)}

\item{rotation_label}{label to use for rotation}

\item{face_label}{labels for x- and z-faces}

\item{palette}{RColorBrewer color palette to use}

\item{pole_label}{label to plot at pole}

\item{n_circle}{number of points to use for drawing Mohr circle}

\item{color_circle}{color of the Mohr circle}

\item{color_lines}{color of lines crossing midpoint of circle}

\item{effective_stress}{if `TRUE`, effective stress label is plotted on the
x-axis. if `FALSE`, total stress}
}
\value{
ggplot object
}
\description{
Function plots a Mohr circle for a rotated stress element and indicates the
pole and the rotation angles
}
\examples{
ggplot_mohrcircle(sigz = 40, sigx = 20, tau = 5, theta = pi/8)
}
