% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_foundations.R
\name{ggplot_pileshaft_poulos}
\alias{ggplot_pileshaft_poulos}
\title{ggplot Poulos's chart for friction along pile in sand}
\usage{
ggplot_pileshaft_poulos(palette = "Set1", phi = NULL, type = NULL, nround = 2)
}
\arguments{
\item{palette}{RColorBrewer pallete for line colors}

\item{phi}{value of angle of internal friction for crosshairs}

\item{type}{pile type for crosshairs. Value should be`Bored`, `Driven` or
`Jacked`}

\item{nround}{number of digits to use for rounding Ks*tan(delta)}
}
\value{
plotly object
}
\description{
Create a ggplot chart for Ks*tan(delta) factors for resistance along pile
shafts in sand, according to Poulos

This function uses data from a digitised plot, so may not be 100% accurate
}
\examples{
#empty chart
ggplot_pileshaft_poulos()

#annotated chart
ggplot_pileshaft_poulos(phi = 36, type = "Driven")
}
