% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_foundations.R
\name{ggplot_piletip_berezantsev}
\alias{ggplot_piletip_berezantsev}
\title{ggplot Berezantsev's chart for pile tip bearing capacity factor}
\usage{
ggplot_piletip_berezantsev(palette = "Set1", phi = NULL, nround = 1)
}
\arguments{
\item{palette}{RColorBrewer pallete for line colors}

\item{phi}{list of angles of internal friction (in degrees) for which to
plot crosshairs and show Nq values}

\item{nround}{number of decimals to use in displayed Nq values in
crosshairs}
}
\value{
ggplot object
}
\description{
Create a ggplot chart for pile tip bearing capacity factor Nq as function
of angle of internal friction for a sandy soil, according to Berezantsev.

This function uses data from a digitised plot, so may not be 100% accurate
}
\examples{
#empty chart
ggplot_piletip_berezantsev()

#annotated chart
ggplot_piletip_berezantsev(phi = 35)
}
