% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasticity.R
\name{ggplot_plasticitychart}
\alias{ggplot_plasticitychart}
\title{Plot plasticity chart}
\usage{
ggplot_plasticitychart(
  wL = NULL,
  Ip = NULL,
  add_legend = TRUE,
  textsize_class = 3,
  textsize_legend = 2.5
)
}
\arguments{
\item{wL}{liquid limits for crosshairs (in \%)}

\item{Ip}{plasticity index for crosshairs (in \%)}

\item{add_legend}{if `TRUE`, return a plot with a legend}

\item{textsize_class}{font size of text in behaviour types}

\item{textsize_legend}{font size of text in legend}
}
\value{
a ggplot object
}
\description{
Function to plot a plasticity chart containing both the A and B lines
}
\examples{
#plot empty chart
ggplot_plasticitychart()

#plot chart with annotation
ggplot_plasticitychart(wL = 60, Ip = 20)
}
