% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robertsons_chart.R
\name{ggplot_robertsonschart}
\alias{ggplot_robertsonschart}
\title{ggplot Robertson's chart}
\usage{
ggplot_robertsonschart(
  Fr = NULL,
  Qtn = NULL,
  group = NULL,
  nFr = 25,
  Fr_lim = c(0.1, 10),
  Qtn_lim = c(1, 1000),
  palette = "Set1",
  alpha = 0.25,
  legend_position = "right"
)
}
\arguments{
\item{Fr}{normalised friction values for crosshairs, in \%}

\item{Qtn}{normalised tip resitance values for crosshairs}

\item{group}{labels for crosshairs}

\item{nFr}{number of points on curved polygon edges. Fr interval is split
linearly}

\item{Fr_lim}{Fr lower and upper limit of plot}

\item{Qtn_lim}{Qtn lower and upper limit of plot}

\item{palette}{colormap to use for coloring different zones}

\item{alpha}{transparancy of soil polygon fills}

\item{legend_position}{position of legend}
}
\value{
ggplot object
}
\description{
Plot an empty Robertson's chart
}
\examples{
#empty chart
ggplot_robertsonschart()

#annotated chart
ggplot_robertsonschart(Fr = c(0.5, 2), Qtn = c(200, 50), group = c("A", "B"))
}
