% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shallow_foundations.R
\name{ggplot_shallowfailuremechanism}
\alias{ggplot_shallowfailuremechanism}
\title{ggplot for block failure mechanism underneath shallow foundation}
\usage{
ggplot_shallowfailuremechanism(
  uB = 0.25,
  phi_deg = 30,
  nwedge = 10,
  symmetrical = FALSE,
  beta_passive_deg = NULL,
  beta_active_deg = NULL,
  hB_foundation = 0.15,
  xlim = c(NA, NA),
  ylim = c(-1, NA),
  fill_soil = "#d3bc5f",
  fill_soil2 = "#bca134",
  color_soil = "#65571d",
  fill_foundation = "black",
  color_foundation = "black",
  size_soil = 0.3,
  label_text = "",
  label_parse = FALSE,
  label_size = 3.5,
  label_nsignif = 4,
  arrow_interval = 0.2,
  arrow_load = 0.4,
  arrow_surcharge = 0.2,
  arrow_size = 0.35,
  arrow_headsize = 0.04,
  color_load = "black",
  color_surcharge = "#6a0dad"
)
}
\arguments{
\item{uB}{vertical displacement of foundation, relative to width B}

\item{phi_deg}{soil angle of internal friction, in degrees}

\item{nwedge}{number of wedges to use in the triangular section}

\item{symmetrical}{if `FALSE`, the bearing capacity failure mechanism is
plotted on one size only. If `TRUE`, failure is symmetrical (both failing
towards the right and the left simultaneously)}

\item{beta_passive_deg}{angle of the passive wedge, in degrees.
Defaults to 45 + phi_deg/4}

\item{beta_active_deg}{angle of the active wedge. in degrees.
Defaults to 45 - phi_deg/4}

\item{hB_foundation}{thickness of the plotted foundation}

\item{xlim}{user-defined x-axis limits}

\item{ylim}{user-defined y-axis limits}

\item{fill_soil}{fill color for soil}

\item{fill_soil2}{fill color for moving soil}

\item{color_soil}{color for outline of soil}

\item{fill_foundation}{fill color for foundation}

\item{color_foundation}{color for outline of foundation}

\item{size_soil}{line thickness of soil polygon outlines}

\item{label_text}{text to plot in upper right hand corner. If set to
`label_text = Nc_auto`, the Nc coefficient is calculated and
displayed}

\item{label_parse}{determines whether the label should be parsed or not}

\item{label_size}{font size of label}

\item{label_nsignif}{= number of significant digits shown for Nc, in case
`label_text = Nc_auto`}

\item{arrow_interval}{distance between load arrows (normalised by foundation
width B)}

\item{arrow_load}{size of load arrow (normalised by foundation
width B). If `arrow_load = 0`, no arrows are plotted}

\item{arrow_surcharge}{size of surcharge arrow (normalised by foundation
width B). If `arrow_surcharge = 0`, no arrows are plotted}

\item{arrow_size}{thickness of arrows}

\item{arrow_headsize}{arrow tip size, in ggplot unit 'npc' (normalised plot
size)}

\item{color_load}{color of load arrow and text}

\item{color_surcharge}{color of surcharge arrows and text}
}
\value{
a ggplot object
}
\description{
Creates a ggplot for shallow bearing capacity mechanism, assuming
blocks sliding across each other (upper bound mechanism)
}
\examples{
#undrained - undisplaced blocks - symmetrical
ggplot_shallowfailuremechanism(uB = 0, phi_deg = 0, symmetrical = TRUE)

#drained - displaced blocks - one-sided mechanism
ggplot_shallowfailuremechanism(uB = 0.4, phi_deg = 30, symmetrical = FALSE)
}
