% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraction_dilation.R
\name{ggplot_simpleshear_dilation}
\alias{ggplot_simpleshear_dilation}
\title{Plot simple shear element with dilating soil}
\usage{
ggplot_simpleshear_dilation(
  u = 1,
  width = 1,
  n = 4,
  nc = 91,
  crop = TRUE,
  marg = 0.02,
  color_water = "#2a7fff",
  color_soil = "#d3bc5f",
  size_water = 1,
  size_soil = 0.5
)
}
\arguments{
\item{u}{relative shear displacement, between 0 (no deformation, dense)
and 1 (max deformation, loose)}

\item{width}{width of the sample relative to its height}

\item{n}{number of grains to use over the sample height}

\item{nc}{number of vertices for plotting circles for grains}

\item{crop}{if `TRUE`, only parts of particles that are present within
the element of soil are plotted. If `FALSE`, also the parts of the
particles that are sticking out are plotted}

\item{marg}{plot margins, relative to the height of the soil sample}

\item{color_water}{color for soil particles}

\item{color_soil}{color for water}

\item{size_water}{line thickness for water/soil element}

\item{size_soil}{line thickness for soil particles}
}
\value{
a ggplot object
}
\description{
Function returns a plot with a simple shear-deformed element filled
with soil particles, starting off in a dense state and dilating
with deformation
}
\examples{
ggplot_simpleshear_dilation(u = 0, n = 3)
ggplot_simpleshear_dilation(u = 1, n = 3)
}
