% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_solutions.R
\name{ggplot_stress_fadum}
\alias{ggplot_stress_fadum}
\title{ggplot Fadum's chart}
\usage{
ggplot_stress_fadum(
  palette = NULL,
  Lz_inf = 1e+06,
  Bz = NULL,
  Lz = NULL,
  nround = 3
)
}
\arguments{
\item{palette}{RColorBrewer color palette to use to more easily differentiate
between lines.}

\item{Lz_inf}{L/z value associated with a strip, i.e L/z --> infinity. Choose
very large}

\item{Bz, Lz}{arrays with values for which to draw crosshairs. If these are
specified, all Fadum's lines are drawn as black, and the annotated crosshairs
are plotted in color}

\item{nround}{number of digits to use in rounding in crosshair labels}
}
\value{
ggplot object
}
\description{
Plot Fadum's chart for the influence factors for a point underneath a
corner of a rectagular uniform load
}
\examples{
#plot empty chart
ggplot_stress_fadum()

#plot with annotation
ggplot_stress_fadum(Bz = 1, Lz = 1)
}
