% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_solutions.R
\name{ggplot_stress_giroud}
\alias{ggplot_stress_giroud}
\title{ggplot Giroud's chart}
\usage{
ggplot_stress_giroud(LB_inf = 1e+06, zB = NULL, LB = NULL, nround = 3)
}
\arguments{
\item{LB_inf}{L/B value associated with a strip, i.e L/B --> infinity. Choose
very large}

\item{zB, LB}{arrays with values for which to draw crosshairs. If these are
specified, all Giroud's lines are drawn as black, and the annotated crosshairs
are plotted in color}

\item{nround}{number of digits to use in rounding in crosshair labels}
}
\value{
ggplot object
}
\description{
Plot Giroud's chart for the influence factors for a point underneath a
corner of a rectagular uniform load
}
\examples{
#plot empty chart
ggplot_stress_giroud()

#plot with annotation
ggplot_stress_giroud(zB = 1, LB = 1)
}
