% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohr_circles.R
\name{ggplot_stresselement}
\alias{ggplot_stresselement}
\title{ggplot to plot rotated stress element}
\usage{
ggplot_stresselement(
  sigz = 40,
  sigx = 20,
  tau = 5,
  theta = 0/180 * pi,
  rotation_label = "theta",
  face_label = c("X", "Z"),
  arrow_offset = 0.1,
  arrow_length = 0.5,
  palette = "Set1",
  fill_soil = "#d3bc5f",
  color_soil = "#65571d",
  coordinate_system = TRUE,
  coordinate_arrow_length = 0.5
)
}
\arguments{
\item{sigz}{normal stress on z-plane (scalar)}

\item{sigx}{normal stress on x-plane (scalar)}

\item{tau}{shear stress}

\item{theta}{rotation, in radians (scalar)}

\item{rotation_label}{label to use for rotation}

\item{face_label}{labels for x- and z-faces}

\item{arrow_offset}{distance between shear arrow and cube}

\item{arrow_length}{maximum length of arrow}

\item{palette}{RColorBrewer color palette to use}

\item{fill_soil}{color of fill of soil cube}

\item{color_soil}{color of outline of soil cube}

\item{coordinate_system}{if `TRUE`, show arrows with the coordinate system
used}

\item{coordinate_arrow_length}{length or coordinate system arrows. Defined
as fraction of `arrow_length`}
}
\value{
ggplot object
}
\description{
Function plots a rotated stress elements and stress arrows. Magnitude of
the arrows scales with magnitude of stress. Normal stress arrows scale by
the major principle stress, shear arrows scale by radius of Mohr circle.

Unit cube of soil has side 1 by 1, and its centre is centred on x=0, y=0
}
\examples{
ggplot_stresselement(sigz = 40, sigx = 20, tau = 5, theta = pi/8)
}
