% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triaxial_testing.R
\name{ggplot_triaxial_mohrcoulomb}
\alias{ggplot_triaxial_mohrcoulomb}
\title{ggplot Mohr-Coulomb failure criterion based on Mohr circles}
\usage{
ggplot_triaxial_mohrcoulomb(
  sig1 = c(75, 120, 160),
  sig3 = c(20, 40, 55),
  drained = TRUE,
  cohesionless = FALSE,
  palette = "Set1",
  nc = 181,
  sig_angle_line = 0.2,
  sig_angle_arrow = 0.18,
  sig_angle_label = 0.19,
  sig_cohesion_arrow = 0.02,
  sig_cohesion_label = 0.03
)
}
\arguments{
\item{sig1}{major principal stresses (array)}

\item{sig3}{minor principal stresses (array)}

\item{drained}{if `TRUE`, stresses are assumed to be effective, and an
cohesional/frictional fit is returned. If `FALSE`, stresses are assumed
to be total stresses, and a undrained shear strength fit is returned}

\item{cohesionless}{if `TRUE`, assume no cohesion when analysing a drained
problem}

\item{palette}{RColorBrewer palette for plotting circles}

\item{nc}{number of points to use to plot a circle}

\item{sig_angle_line}{relative x-length of horizontal line for angle plot}

\item{sig_angle_arrow}{relative x-position of phi arrow}

\item{sig_angle_label}{relative x-position of phi label}

\item{sig_cohesion_arrow}{relative x-position of cohesion arrow}

\item{sig_cohesion_label}{relative x-position of cohesion label}
}
\value{
ggplot object
}
\description{
Plot an drained or undrained mohr-coulomb fit based on measured
principal stresses.

Mohr-Coulomb fits are obtained by minimising the sum of square distances
between failure surface and Mohr circles
}
\examples{
#drained example - cohesional/frictional
sig1 = c(100, 130, 160)
sig3 = c(50, 70, 85)
ggplot_triaxial_mohrcoulomb(sig1, sig3, drained = TRUE)

#drained example - cohesionless
sig1 = c(100, 130, 160)
sig3 = c(50, 70, 85)
ggplot_triaxial_mohrcoulomb(sig1, sig3, drained = TRUE, cohesionless = TRUE)

#undrained example
sig1 = c(100, 130, 160)
sig3 = c(50, 75, 102)
ggplot_triaxial_mohrcoulomb(sig1, sig3, drained = FALSE)
}
