% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_profiles_horizontal.R
\name{horizontal_stress_profile}
\alias{horizontal_stress_profile}
\title{Calculate profiles of horizontal effective stresses}
\usage{
horizontal_stress_profile(
  z_soil = c(0, 4, 6),
  z_watertable = 4,
  q = 0,
  gamma_b = c(18, 20, 22),
  z_soil0 = NULL,
  z_watertable0 = NULL,
  q0 = NULL,
  gamma_b0 = NULL,
  z_max = 10,
  z_interval = 0.25,
  gamma_w = 10,
  phi_deg = c(30, 20, 25),
  description = "sand"
)
}
\arguments{
\item{z_soil}{depth of tops of current soil layer,}

\item{z_watertable}{current water table level,}

\item{q}{current surcharge,}

\item{gamma_b}{current bulk unit weights}

\item{z_soil0}{historic depths of tops of soil layers}

\item{z_watertable0}{historic water table}

\item{q0}{historic surcharge}

\item{gamma_b0}{historic bulk unit weights}

\item{z_max}{maximum depth to consider}

\item{z_interval}{depth interval at which to calculate stresses}

\item{gamma_w}{unit weight of water}

\item{phi_deg}{angle of internal friction for current soil layers
(in degrees)}

\item{description}{type of current soil layers, either `sand` or `clay`.
used to determine K0}
}
\value{
a tibble with many values, e.g. `sigma'_v` and `sigma'_v0` for
  current and historic vertical effective stresses, `OCR` for
  overconsolidation ratio, `K0NC` and `K0` for normally consolidated and
  current coefficient of lateral earth pressure, and `sigma'_h` and
  `sigma'_hNC` for current and normally consolidated horizontal effective
  stress
}
\description{
Function takes two stress profiles, one corresponding with the largest
historical vertical effective stress and one with the current stresses,
and generates profiles with horizontal stresses based on soil properties
}
\examples{
horizontal_stress_profile()
}
