% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_foundations.R
\name{hyperbolicmethod_settlement}
\alias{hyperbolicmethod_settlement}
\title{Calculate pile displacements using the hyperbolic method}
\usage{
hyperbolicmethod_settlement(Q, Qb, Qs, Eb, Ds, Db = NULL, Ms = 0.001)
}
\arguments{
\item{Q}{vector of loads applied to pile}

\item{Qb}{pile base capacity}

\item{Qs}{pile shaft capacity}

\item{Eb}{soil Young's modulus underneath the pile tip}

\item{Ds}{pile shaft diameter}

\item{Db}{pile base diameter. If not defined, this is assumed to be equal
to the shaft diameter}

\item{Ms}{empirical factor, assumed as Ms = 0.001 by default}
}
\description{
Calculates the settlement of a rigid pile using Fleming's (1992)
hyperbolic method for pile displacements.
}
\examples{
# Input parameters
Qs <- 1000  # [kN]
Qb <- 1500  # [kN]
Q <- seq(0, Qs + Qb, l = 101)  # [kN]
Eb <- 20*1e3  # [kPa]
Ds <- 0.5  # [m]

# Calculate settlement
s <- hyperbolicmethod_settlement(Q, Qb, Qs, Eb, Ds)  # [m]

# plot
plot(Q, s)
}
