% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robertsons_chart.R
\name{IcFr2Qtn}
\alias{IcFr2Qtn}
\title{Calculate Qtn from known Ic and Fr, in zone 2 - 7}
\usage{
IcFr2Qtn(Ic, Fr)
}
\arguments{
\item{Ic}{Robertson's Ic index}

\item{Fr}{normalised friction ratio, in \%}
}
\value{
normalised tip resistance Qtn
}
\description{
Calculate Ic from known Ic and Fr, in zone 2 - 7 in Robertson's
chart
}
