% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robertsons_chart.R
\name{IcQtn2Fr}
\alias{IcQtn2Fr}
\title{Calculate Fr from known Qtn and Ic, in zone 2 - 7}
\usage{
IcQtn2Fr(Ic, Qtn)
}
\arguments{
\item{Ic}{Robertson's Ic index}

\item{Qtn}{normalised tip resistance}
}
\value{
normalised friction ratio, in \%
}
\description{
Calculate Fr from known Qtn and Ic, in zone 2 - 7 in Robertson's
chart
}
