% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_differences.R
\name{index_edge}
\alias{index_edge}
\title{Get indices of nodes near edge of grid}
\usage{
index_edge(nx, ny, edge, real_only = FALSE, i0 = 0, offset = 0, ...)
}
\arguments{
\item{nx, ny}{number of real nodes in x and y-direction}

\item{edge}{edge number, numbered clockwise. `1` for left (negative x-face),
`2` for top (positive y-face), `3` for right (positive x-face) and `4`
for bottom (negative y-face)}

\item{real_only}{if `TRUE`, only real nodes are numbered. if `FALSE`, all
ghost nodes are included in the numbering as well}

\item{i0}{node starting number (index of first node minus 1)}

\item{offset}{row (in case `edge = 2` or `edge = 4`, or column (in case
`edge = 1` or `edge = 3`) offset from edge}

\item{...}{additional optional arguments}
}
\value{
array of indices of ghost nodes on selected edge
}
\description{
Get an array of the indices of nodes near an edge of a rectangular grid.
For node numbering, see function `findiff_sparse_entries()`.
}
\examples{
#real nodes + ghost nodes
index_edge(4, 3, 1)
index_edge(4, 3, 2, offset = 1)

#real nodes only
index_edge(4, 3, 1, real_only = TRUE)
index_edge(4, 3, 2, offset = 1, real_only = TRUE)
}
