% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R
\name{index_ghost}
\alias{index_ghost}
\title{Get indices of ghost nodes on an edge}
\usage{
index_ghost(nx, ny, edge, i0 = 0)
}
\arguments{
\item{nx, ny}{number of real nodes in x and y-direction}

\item{edge}{edge number, numbered clockwise. `1` for left (negative x-face),
`2` for top (positive y-face), `3` for right (positive x-face) and `4`
for bottom (negative y-face)}

\item{i0}{node starting number.}
}
\value{
array of indices of ghost nodes on selected edge
}
\description{
Get an array of the indices of all ghost nodes on a side of a rectangular
grid of (real) nodes. Nodes are number per row --> per column. Ghost nodes
are placed outside the edges of all sides of the rectangular grid
}
