% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_differences.R, R/flow_nets.R
\name{index_neighbour}
\alias{index_neighbour}
\title{Get indices of nodes bordering nodes}
\usage{
index_neighbour(i, nx, ny, edge, i0 = 0)

index_neighbour(i, nx, ny, edge, i0 = 0)
}
\arguments{
\item{i}{array with node indices}

\item{nx, ny}{number of nodes in x and y-direction}

\item{edge}{side on which to look, numbered clockwise. `1` for left
(negative x-face), `2` for top (positive y-face), `3` for right
(positive x-face) and `4` for bottom (negative y-face)}

\item{i0}{node starting number (index of first node minus 1)}

\item{offset}{number of nodes to move in specific direction}
}
\value{
an array with node indices for neighbours

an array with node indices for neighbours
}
\description{
Get the indices of neighbouring nodes in a rectangular grid with nodes.
For node numbering, see function `findiff_sparse_entries()`

Get the indices of neighbouring nodes in a grid with real and ghost nodes.
This function inherently assumes those nodes exists. Make sure you don't
request any nodes outside the grid.
}
