% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_finitedifferences.R
\name{index_offset}
\alias{index_offset}
\title{Find indices of neighbouring nodes}
\usage{
index_offset(i, nx, ny, dix = 0, diy = 0, i0 = 0, real_only = FALSE, ...)
}
\arguments{
\item{i}{array with node indices}

\item{nx}{number of real nodes in x and y directions}

\item{ny}{number of real nodes in x and y directions}

\item{dix, diy}{index offsets in x and y-directions}

\item{i0}{node index offset}

\item{real_only}{if `TRUE`, only real nodes are numbered and ghost nodes
are ignored in numbering}

\item{...}{additional arguments to pass}
}
\value{
a list of indices of neightbouring nodes
}
\description{
Find indices of neighbouring nodes in a rectangular grid with or without
ghost nodes. Nodes are numbered row-wise. Function assumes that these
nodes exist, e.g. no check is conducted to see if this is so...
}
\examples{
nx <- 4
ny <- 3
i <- c(1, 11, 20)
index_offset(i, nx, ny, dix = 1)
index_offset(i, nx, ny, diy = 1)
}
