% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_finitedifferences.R
\name{index_real}
\alias{index_real}
\title{Get indices of all real nodes}
\usage{
index_real(nx, ny, i0 = 0, real_only = FALSE, ...)
}
\arguments{
\item{nx}{number of real nodes in x and y-direction}

\item{ny}{number of real nodes in x and y-direction}

\item{i0}{node starting number (index of first node minus 1)}

\item{real_only}{if `TRUE`, only real nodes are numbered. if `FALSE`, all
ghost nodes are included in the numbering as well}

\item{...}{additional optional arguments}
}
\value{
array of indices for all real nodes (non-ghost nodes)
}
\description{
Get an array of the indices of all real nodes in a rectangular grid. For
node numbering, see function `findiff_sparse_elements()`
}
