% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R
\name{index_real_edge}
\alias{index_real_edge}
\title{Get index of real nodes on an edge}
\usage{
index_real_edge(nx, ny, edge, i0 = 0, real_only = FALSE, ...)
}
\arguments{
\item{nx}{number of real nodes in x and y-direction}

\item{ny}{number of real nodes in x and y-direction}

\item{edge}{edge number, numbered clockwise. `1` for left (negative x-face),
`2` for top (positive y-face), `3` for right (positive x-face) and `4`
for bottom (negative y-face)}

\item{i0}{node starting number.}

\item{real_only}{if `FALSE`, ghost nodes are assumed to be included in the
numbering of indices. If `TRUE`, all ghost nodes are assumed to have been
removed, and only real points are numbered}

\item{...}{additional arguments, in case the function is passed to
`purrr::pmap()`}
}
\value{
array with indices of real nodes on edge
}
\description{
Get indices of real nodes lying on an edge. Indices are numbered row-first
(x first).
}
