% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_solutions.R
\name{influencefactor_cornerrectangle}
\alias{influencefactor_cornerrectangle}
\title{Influence factor for stress under corner of rectangular uniform load}
\usage{
influencefactor_cornerrectangle(xz, yz, direction = "z")
}
\arguments{
\item{xz}{x-coordinate, normalised by depth z}

\item{yz}{y-coordinate, normalised by depth z}

\item{direction}{direction of stress increase. `z` for vertical}
}
\value{
array with influence factors
}
\description{
Calculates the influence factor for a point underneath the corner of a
rectangular uniform load
}
