% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_solutions.R
\name{influencefactor_triangularstrip}
\alias{influencefactor_triangularstrip}
\title{Influence factor for stress under triangular load}
\usage{
influencefactor_triangularstrip(xB, zB, direction = "z", nu = 0.3)
}
\arguments{
\item{xB}{x-coordinate, normalised by width of strip (array). `xB = 0` is
located where the triangle height is 0, and `xB = 1` where the triangular
load is at it maximum (just before it drops to zero after `xB > 1`)}

\item{zB}{depth, normalised by width of strip (array)}

\item{direction}{direction of stress increase. `z` for vertical, `x` for
horizontal away from strip, `y` for horizontal in direction of strip, and
`xz` for shear stress}

\item{nu}{Posisson's ratio}
}
\value{
array with influence factors
}
\description{
Calculates the influence factor for the total stress underneath
a strip load, based on the position in normalised coordinates
}
