% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_solutions.R
\name{influencefactor_uniformstrip}
\alias{influencefactor_uniformstrip}
\title{Influence factor for stress under strip load}
\usage{
influencefactor_uniformstrip(xB, zB, direction = "z", nu = 0.3)
}
\arguments{
\item{xB}{x-coordinate, normalised by width of strip (array)}

\item{zB}{depth, normalised by width of strip (array)}

\item{direction}{direction of stress increase. `z` for vertical, `x` for
horizontal away from strip, `y` for horizontal in direction of strip, and
`xz` for shear stress}

\item{nu}{Posisson's ratio}
}
\value{
array with influence factors
}
\description{
Calculates the influence factor for the total stress underneath
a strip load, based on the position in normalised coordinates
}
