% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_plotting.R
\name{is_grid_rectangular}
\alias{is_grid_rectangular}
\title{Check if series of gridded domains is rectangular}
\usage{
is_grid_rectangular(dp)
}
\arguments{
\item{dp}{tibble with flow net solution. Should contain fields `x` and `y`
for positions of the nodes}
}
\value{
`TRUE` is the grid is rectangular, `FALSE` is not
}
\description{
Check if a flownet solution is given on a rectangular grid, i.e. nodes
are organised in horizontal rows and vertical columns. The distances
between rows and columns may vary.

For each domain in the grid, the function checks if the number of nodes
is equal to the product of unique x and y-positions. If this is true for
all domains, the grid is regular and the function returns `TRUE`.
}
\examples{
#rectangular grid
df <- flownet_geometry_rectangular()
dp <- flownet_solve_rectangular(df)
is_grid_rectangular(dp)

#quadrilateral grid
df <- flownet_geometry_quadrilateral()
dp <- flownet_solve_quadrilateral(df)
is_grid_rectangular(dp)
}
