% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_rectangular.R
\name{ispresent_4wayintersections}
\alias{ispresent_4wayintersections}
\title{Check for 4-intersections}
\usage{
ispresent_4wayintersections(df)
}
\arguments{
\item{df}{list with dataframes with problem description}
}
\value{
boolean. `TRUE` if 4-way intersections exist
}
\description{
Solver technique seems to get stuck when four domains are touching.
The point on the 4-way intersection may throw a numerical problem. This
function checks if such points exist.
}
\examples{
df <- flownet_geometry_rectangular()
ispresent_4wayintersections(df)
}
