% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geometry.R
\name{lseq}
\alias{lseq}
\title{Distribute points evenly on log-space interval}
\usage{
lseq(from, to, ...)
}
\arguments{
\item{from}{start of interval}

\item{to}{end of interval}

\item{...}{additional arguments to `seq()` function}
}
\value{
array with values
}
\description{
Equivalent of R's `seq()` function, but then on a log scale.
Similar to logscape function in Matlab
}
