% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_quadrilateral.R
\name{matrix_sparse_head_quadrilateral}
\alias{matrix_sparse_head_quadrilateral}
\title{Obtain sparse linear system for solving hydraulic head (quadrilateral)}
\usage{
matrix_sparse_head_quadrilateral(df)
}
\arguments{
\item{df}{list with tibbles describing the problem, see function
`flownet_geometry_quadrilateral()` for more information}
}
\value{
list with two elements. `mat` contains a tibble with all sparse
  matrix elements (`row`, `col`, `val`), and `lhs` contains a vector
  with all left-hand side of equation values
}
\description{
Take the geometry and boundary conditions of a flownet problem consisting
of connected quadrilateral domains, and returns the matrix and vector of
the linear finite difference problem
}
\examples{
df <- flownet_geometry_quadrilateral()
matrix_sparse_head_quadrilateral(df)
}
