% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc_calculations.R
\name{mcc_matrix_elastoplastic}
\alias{mcc_matrix_elastoplastic}
\title{Generate MCC elasto-plastic matrix elements}
\usage{
mcc_matrix_elastoplastic(p, q, pc, v0, M, kappa, lambda, nu)
}
\arguments{
\item{p, q}{current stress invariants}

\item{pc}{current preconsolidation pressure}

\item{v0}{initial specific volume}

\item{M}{M-parameter}

\item{kappa, lambda}{MCC compression parameters}

\item{nu}{Poisson's ratio}
}
\value{
array with four stiffness elements [M11, M12, M21, M22]
}
\description{
Returns the matrix elements of the elasto-plastic stiffness matrix
according to the modified cam-clay model, in terms of isotropic and
deviatoric stress invariants p and q, and volumetric and deviatoric
strain invariants epsv and epq (epsv is assumed positive in
compression)

dp = M11*depsv + M12*depsq
dq = M21*depsv + M22*depsq
}
