% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc_calculations.R
\name{mcc_triaxcomp_drained_plastic}
\alias{mcc_triaxcomp_drained_plastic}
\title{Solve incremental plastic MCC response - drained triaxial compression}
\usage{
mcc_triaxcomp_drained_plastic(de1, p, q, pc, v0, M, kappa, lambda, nu)
}
\arguments{
\item{de1}{axial strain increment}

\item{p}{current stress invariants}

\item{q}{current stress invariants}

\item{pc}{current preconsolidation pressure}

\item{v0}{initial specific volume}

\item{M}{M-parameter}

\item{kappa}{MCC compression parameters}

\item{lambda}{MCC compression parameters}

\item{nu}{Poisson's ratio}
}
\value{
array with increment in radial strain and increment in axial stress
}
\description{
Solve the incremental response for a drained triaxial compression
test according to MCC theory, when behaviour is plastic
}
