% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc_plot.R
\name{mcc_yieldsurface_pq}
\alias{mcc_yieldsurface_pq}
\title{MCC yield surface coordinates (positive side only)}
\usage{
mcc_yieldsurface_pq(pc, M, t = seq(0, pi, l = 101))
}
\arguments{
\item{pc}{preconsolidation pressure}

\item{M}{MCC M-parameters}

\item{t}{array with angles at which to plot point}
}
\value{
tibble with `p` and `q`
}
\description{
Create p and q positions for a MCC yield surface
}
