% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_rectangular.R
\name{nodal_coordinates_real_rectangular}
\alias{nodal_coordinates_real_rectangular}
\title{Get coordinates of all real nodes in the finite difference grid}
\usage{
nodal_coordinates_real_rectangular(
  nx,
  ny,
  x0 = 0,
  y0 = 0,
  x1 = 1,
  y1 = 1,
  domain = NULL,
  ...
)
}
\arguments{
\item{nx, ny}{number of real nodes on the grid}

\item{x0, x1, y0, y1}{x and y-positions of domain edges}

\item{domain}{array with domain identifiers}

\item{...}{additional named arguments to pass to function}
}
\value{
a tibble with fields `x` and `y` for positions, and `id` to indicate
  which grid the point belongs to
}
\description{
Function obtains the x,y positions of all real nodes in a rectangular
finite difference grid
}
\examples{
df <- nodal_coordinates_real_rectangular(
  nx = c(4, 5),
  ny = c(3, 4),
  x0 = c(0, 4),
  y0 = c(0, 4),
  x1 = c(2, 7),
  y1 = c(3, 5)
)
plot(df$x, df$y, "b")
}
