% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_differences.R
\name{nodes_total}
\alias{nodes_total}
\title{Determine total number of nodes}
\usage{
nodes_total(nx, ny, real_only = FALSE, ...)
}
\arguments{
\item{nx}{number of real nodes in x and y-direction}

\item{ny}{number of real nodes in x and y-direction}

\item{real_only}{if `TRUE`, only real nodes are numbered. if `FALSE`, all
ghost nodes are included in the numbering as well}

\item{...}{additional optional arguments}
}
\value{
number of nodes (scalar)
}
\description{
Determine the total number of nodes on a grid of nodes, including any
ghost nodes outside all four sides of the grid
}
