% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oedometer.R
\name{oedometer_create_data}
\alias{oedometer_create_data}
\title{Create oedometer time-sample height data for single load step}
\usage{
oedometer_create_data(
  t = NULL,
  h0 = 20 * 0.001,
  dsigmav = 20000 * 1e+06,
  mv = 5e-06,
  cv = 1e-08,
  Calpha = 0.01,
  Tvalpha = 10^0.2
)
}
\arguments{
\item{t}{array with time steps. If not defined, a wide range is assumed}

\item{h0}{initial sample height [m]}

\item{dsigmav}{increase in vertical stress [Pa]}

\item{mv}{coefficient of compressibility [m^2/N]}

\item{cv}{coefficient of consolidation [m^2/s]}

\item{Calpha}{creep strain per log10-cycle of time. Creep is only added
after primary consolidation is (almost) finished. A small smoothing
is applied to smoothen the sudden jump in gradient of settlement}

\item{Tvalpha}{value of normalised strain at which creep stats to be
included}
}
\value{
tibble with field for time (`t`, in seconds) and current sample
  height (`h`, in meters)
}
\description{
Create some time - sample height oedometer data, given known soil
properties, for a single load step
}
\examples{
#get data
df <- oedometer_create_data()

#plot data - log scale for time
ggplot2::ggplot() +
  ggplot2::geom_line(
    data = df,
    ggplot2::aes(x = t, y = h)
  ) +
  ggplot2::scale_x_log10()
}
