% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shallow_foundations.R
\name{plotly_bearingcapacityfactors}
\alias{plotly_bearingcapacityfactors}
\title{Plotly chart with foundation bearing capacity factors}
\usage{
plotly_bearingcapacityfactors(
  phi_deg = seq(0, 50, 1),
  ylim = c(1, 1000),
  nround = 1,
  palette = "Set1"
)
}
\arguments{
\item{phi_deg}{array with all values for the angle of internal friction,
in degrees (array)}

\item{ylim}{2-value array with y-axis limits}

\item{nround}{number of decimals to round angle and N values to}

\item{palette}{RCOlorBrewer color palette to use for line colors}
}
\value{
plotly object
}
\description{
Creates an interactive plotly chart with the shallow foundation bearing
capacity factors, varying with the angle of internal friction, according
to EC7
}
\examples{
plotly_bearingcapacityfactors()
}
