% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_foundations.R
\name{plotly_pileshaft_poulos}
\alias{plotly_pileshaft_poulos}
\title{plotly Poulos's chart for friction along pile in sand}
\usage{
plotly_pileshaft_poulos(palette = "Set1", nround = 2)
}
\arguments{
\item{palette}{RColorBrewer pallete for line colors}

\item{nround}{number of decimals to round values in hoverlabels to}
}
\value{
plotly object
}
\description{
Create a plotly chart for Ks*tan(delta) factors for resistance along pile
shafts in sand, according to Poulos

This function uses data from a digitised plot, so may not be 100% accurate
}
\examples{
plotly_pileshaft_poulos()
}
