% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile_foundations.R
\name{plotly_piletip_berezantsev}
\alias{plotly_piletip_berezantsev}
\title{plotly Berezantsev's chart for pile tip bearing capacity factor}
\usage{
plotly_piletip_berezantsev(palette = "Set1")
}
\arguments{
\item{palette}{RColorBrewer pallete for line colors}
}
\value{
plotly object
}
\description{
Create a plotly chart for pile tip bearing capacity factor Nq as function
of angle of internal friction for a sandy soil, according to Berezantsev.

This function uses data from a digitised plot, so may not be 100% accurate
}
\examples{
plotly_piletip_berezantsev()
}
