% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profiles_horizontal.R
\name{plotly_soilstressprofile_horizontal}
\alias{plotly_soilstressprofile_horizontal}
\title{Plotly calculation process of current horizontal effective stresses}
\usage{
plotly_soilstressprofile_horizontal(
  z_soil = c(0, 4),
  z_watertable = 4,
  q = 0,
  gamma_b = c(18, 20),
  z_soil0 = NULL,
  z_watertable0 = NULL,
  q0 = 100,
  gamma_b0 = NULL,
  z_max = 10,
  z_interval = 0.1,
  gamma_w = 10,
  phi_deg = c(35, 15),
  description = c("sand", "clay"),
  ylab = "Depth [m]",
  mode = "lines",
  nrow = 2,
  width = 800,
  height = 600
)
}
\arguments{
\item{z_soil}{depth of tops of current soil layer,}

\item{z_watertable}{current water table level,}

\item{q}{current surcharge,}

\item{gamma_b}{current bulk unit weights}

\item{z_soil0}{historic depths of tops of soil layers}

\item{z_watertable0}{historic water table}

\item{q0}{historic surcharge}

\item{gamma_b0}{historic bulk unit weights}

\item{z_max}{maximum depth to consider}

\item{z_interval}{depth interval at which to calculate stresses}

\item{gamma_w}{unit weight of water}

\item{phi_deg}{angle of internal friction for current soil layers
(in degrees)}

\item{description}{type of current soil layers, either `sand` or `clay`.
used to determine K0}

\item{ylab}{depth axis label}

\item{mode}{plotly trace mode, e.g. `lines`}

\item{nrow}{number of rows for subplots}

\item{width}{width of plot}

\item{height}{height of plot}
}
\value{
a plotly object
}
\description{
Function takes data for two soil profiles (current and historical when
effective stress was largests) and draws soil profiles, profiles with
vertical effective stress, OCR, K0 and horizontal effective stress with
depth
}
\examples{
#removal of surcharge
plotly_soilstressprofile_horizontal(q = 0, q0 = 100)

#raising of the water table
plotly_soilstressprofile_horizontal(q = 0, q0 = 0, z_watertable = 0, z_watertable0 = 6)
}
