% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_profiles_vertical.R
\name{plotly_soilstressprofile_vertical}
\alias{plotly_soilstressprofile_vertical}
\title{Plotly soil profile and vertical stress profile side-by-side}
\usage{
plotly_soilstressprofile_vertical(
  z_soil,
  z_watertable = 0,
  z_max = 10,
  z_interval = NULL,
  q = 0,
  gamma_b = 18,
  traces = c("sigma_v", "u", "sigma'_v"),
  labelmode = "stressonly",
  width = 600,
  height = 300
)
}
\arguments{
\item{z_soil}{depth of top of each layer}

\item{z_watertable}{depth of the water table}

\item{z_max}{maximum depth of interest}

\item{z_interval}{additional interval at which to calculate stresses}

\item{q}{surcharge at soil surface}

\item{gamma_b}{unit weight of soil, in kN/m3. Should be a scalar or a
vector with the same length as `z_soil`}

\item{traces}{array with traces to be plotted. May contain `sigma_v` for
the total vertical stress, `u` for the pore water pressure and `sigma'_v`
for the vertical effective stress}

\item{labelmode}{setting controlling what hoverlabels to plot. If
`stressonly`, only the values of stresses are displaced. If
`calculations`, it is also shown how to calculate the stress based on the
stress value in other points}

\item{width}{width of the output plot}

\item{height}{height of the output plot}
}
\value{
a plotly object
}
\description{
Creates a plotly plot with the soil profile on the left-hand side,
and the stress profiles on the right-hand side
}
\examples{
#dry soil - total stress only
plotly_soilstressprofile_vertical(
  z_soil = 0,
  z_watertable = 20,
  z_max = 6,
  z_interval = 1,
  traces = c("sigma_v")
)

#partly saturated soil
plotly_soilstressprofile_vertical(
  z_soil = c(0, 4, 8),
  z_watertable = 2,
  z_max = 10,
  q = 20,
  gamma_b = c(18, 20, 15),
  traces = c("sigma_v", "u", "sigma'_v"),
  labelmode = "calculations"
)
}
