% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_profiles_horizontal.R
\name{plotly_stressprofile_horizontal_ocr}
\alias{plotly_stressprofile_horizontal_ocr}
\title{Plotly OCR trace}
\usage{
plotly_stressprofile_horizontal_ocr(
  z,
  OCR,
  ylab = "Depth [m]",
  xlim = c(0, NA),
  ylim = c(NA, NA),
  mode = "lines",
  line_colour = c("#FF0000", "#707070"),
  line_type = c("solid", "dash"),
  line_name = c("Current", "Normally consolidated"),
  line_width = 1.5,
  nround = 2
)
}
\arguments{
\item{z}{array width depths}

\item{OCR}{array with overconsolidation ratio's for each depth `z`}

\item{ylab}{depth axis label}

\item{xlim}{manually override stress axis limits}

\item{ylim}{manually override depth axis limits}

\item{mode}{plotly plotting mode, e.g. `lines` or `lines+markers`}

\item{line_colour}{array with two colors}

\item{line_type}{array with two plotly line styles}

\item{line_name}{names for the two traces}

\item{line_width}{line width}

\item{nround}{number of digits for rounding hoverlabels}
}
\value{
a plotly object
}
\description{
Plots two traces of vertical effective stress in a single plotly plot
}
\examples{
ds <- horizontal_stress_profile(q0 = 50)
plotly_stressprofile_horizontal_ocr(
  ds$z,
  ds$OCR
)
}
