% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profiles_horizontal.R
\name{plotly_stressprofile_horizontal_sigmav}
\alias{plotly_stressprofile_horizontal_sigmav}
\title{Plotly historic and current vertical stress traces}
\usage{
plotly_stressprofile_horizontal_sigmav(
  z,
  sigma_v0,
  sigma_v,
  ylab = "Depth [m]",
  xlim = c(0, NA),
  ylim = c(NA, NA),
  mode = "lines",
  line_colour = c("#ff0000", "#ff0000"),
  line_type = c("dash", "solid"),
  line_name = c("Historic", "Current"),
  line_width = 1.5,
  nround = 2
)
}
\arguments{
\item{z}{array width depths}

\item{sigma_v0}{array with previously largests vertical effective stress
at depths `z`}

\item{sigma_v}{array with current vertical effective stress at depths `z`}

\item{ylab}{depth axis label}

\item{xlim}{manually override stress axis limits}

\item{ylim}{manually override depth axis limits}

\item{mode}{plotly plotting mode, e.g. `lines` or `lines+markers`}

\item{line_colour}{array with two colors}

\item{line_type}{array with two plotly line styles}

\item{line_name}{names for the two traces}

\item{line_width}{line width}

\item{nround}{number of digits for rounding hoverlabels}
}
\value{
a plotly object
}
\description{
Plots two traces of vertical effective stress in a single plotly plot
}
\examples{
ds <- horizontal_stress_profile(q0 = 50)
plotly_stressprofile_horizontal_sigmav(
  ds$z,
  ds$`sigma'_v0`,
  ds$`sigma'_v`
)
}
