% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_quadrilateral.R,
%   R/flownet_quadrilateral_old.R
\name{poissons_eq_quadrilateral}
\alias{poissons_eq_quadrilateral}
\title{Get sparse Poisson's matrix elements for quadrilateral domain}
\usage{
poissons_eq_quadrilateral(
  nx,
  ny,
  kx,
  ky,
  x1,
  x2,
  x3,
  x4,
  y1,
  y2,
  y3,
  y4,
  i0 = 0,
  ...
)

poissons_eq_quadrilateral(
  nx,
  ny,
  kx,
  ky,
  x1,
  x2,
  x3,
  x4,
  y1,
  y2,
  y3,
  y4,
  i0 = 0,
  ...
)
}
\arguments{
\item{nx, ny}{number of real nodes in a and b-directions}

\item{kx, ky}{horizontal and vertical permeability in domain}

\item{x1, x2, x3, x4, y1, y2, y3, y4}{x and y positions of the corner points of the
quadrilateral. Corners are numbered in clockwise order.}

\item{i0}{index offset for nodes}

\item{...}{extra arguments to pass}
}
\value{
a tibble with rows (`row`), columns (`col`) and values (`val`)
  of non-zero elements in the sparse matrix

a tibble with rows (`row`), columns (`col`) and values (`val`)
  of non-zero elements in the sparse matrix
}
\description{
Function generates the sparse matrix elements for the Poisson's
equation in all real nodes, for a single quadrilateral domain

Function generates the sparse matrix elements for the Poisson's
equation in all real nodes, for a single quadrilateral domain
}
\examples{
nx <- 20
ny <- 10
kx <- 1
ky <- 1
x1 <- 0
x2 <- 0
x4 <- 1
x3 <- 1
y1 <- 0
y2 <- 1
y3 <- 2
y4 <- 0
poissons_eq_quadrilateral(nx, ny, kx, ky, x1, x2, x3, x4, y1, y2, y3, y4)
nx <- 20
ny <- 10
kx <- 1
ky <- 1
x1 <- 0
x2 <- 0
x4 <- 1
x3 <- 1
y1 <- 0
y2 <- 1
y3 <- 2
y4 <- 0
poissons_eq_quadrilateral(nx, ny, kx, ky, x1, x2, x3, x4, y1, y2, y3, y4)
}
