% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geometry.R
\name{polygon_area}
\alias{polygon_area}
\title{Get cross-sectional area of polygon}
\usage{
polygon_area(x, y)
}
\arguments{
\item{x, y}{arrays with x and y coordinates of the polygon}
}
\value{
list with x and y positions
}
\description{
Calculate the cross-sectional area of a polygon based on the coordinates
of the polygon
}
