% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_geometry.R
\name{polygon_centroid}
\alias{polygon_centroid}
\title{Get position of centroid of polygon}
\usage{
polygon_centroid(x, y)
}
\arguments{
\item{x, y}{arrays with x and y coordinates of the polygon}
}
\value{
list with x and y positions
}
\description{
Calculate the centroid position of a polygon based on the coordinates
of the polygon
}
