% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_quadrilateral.R,
%   R/flownet_quadrilateral_old.R
\name{position_xy_quadrilateral}
\alias{position_xy_quadrilateral}
\title{Get x,y positions and derivatives of real nodes in quadrilateral grid}
\usage{
position_xy_quadrilateral(a, b, x1, x2, x3, x4, y1, y2, y3, y4, ...)

position_xy_quadrilateral(a, b, x1, x2, x3, x4, y1, y2, y3, y4, ...)
}
\arguments{
\item{a, b}{arrays with quadrilateral positions}

\item{x1, x2, x3, x4, y1, y2, y3, y4}{x and y positions of the corner points of the
quadrilateral. Corners are numbered in clockwise order.}

\item{...}{additional arguments}
}
\value{
a tibble with fields for position `x` and `y`, as well as 1st order
  derivatives (`x_a`, `x_b`, `y_a`, `y_b`) and second order derivatives
  (`x_aa`, `x_bb`, `x_ab`, `y_aa`, `y_bb`, `y_ab`)

a tibble with fields for position `x` and `y`, as well as 1st order
  derivatives (`x_a`, `x_b`, `y_a`, `y_b`) and second order derivatives
  (`x_aa`, `x_bb`, `x_ab`, `y_aa`, `y_bb`, `y_ab`)
}
\description{
Function calculates the x and y positions of nodes in a quadrilateral grid
based on their a and b positions. The derivatives of x and y with respect
to a and b are also returned

Function calculates the x and y positions of nodes in a quadrilateral grid
based on their a and b positions. The derivatives of x and y with respect
to a and b are also returned
}
\examples{
a <- c(0, 0.5, 1)
b <- c(0, 0.5, 1)
x1 <- 0
x2 <- 0
x3 <- 1
x4 <- 1
y1 <- 0
y2 <- 1
y3 <- 2
y4 <- 0
position_xy_quadrilateral(a, b, x1, x2, x3, x4, y1, y2, y3, y4)
a <- c(0, 0.5, 1)
b <- c(0, 0.5, 1)
x1 <- 0
x2 <- 0
x3 <- 1
x4 <- 1
y1 <- 0
y2 <- 1
y3 <- 2
y4 <- 0
position_xy_quadrilateral(a, b, x1, x2, x3, x4, y1, y2, y3, y4)
}
