% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_quadrilateral.R
\name{positions_real_quadrilateral}
\alias{positions_real_quadrilateral}
\title{Get positions of all real nodes in quadrilateral domains}
\usage{
positions_real_quadrilateral(
  nx,
  ny,
  x1,
  x2,
  x3,
  x4,
  y1,
  y2,
  y3,
  y4,
  domain = 1,
  i0 = 0,
  ...
)
}
\arguments{
\item{nx, ny}{number of real nodes in x and y directions}

\item{x1, x2, x3, x4, y1, y2, y3, y4}{x and y positions of the corner points of the
quadrilateral. Corners are numbered in clockwise order.}

\item{domain}{domain identifier}

\item{i0}{optional index numbering offset}

\item{...}{additional arguments to pass}
}
\value{
a tibble with positions, e.g. `a`, `b`, `x`, `y` as well as
  1st and 2nd order derivatives of positions (e.g. `x_a`, `y_bb`)
}
\description{
Generate positions of all real nodes in a quadrilateral domain, both in
terms of normalised positions a and b, and actual positions x and y
}
\examples{
nx <- 4
ny <- 3
x1 <- 0
x2 <- 0
x3 <- 1
x4 <- 1
y1 <- 0
y2 <- 1
y3 <- 2
y4 <- 0
positions_real_quadrilateral(nx, ny, x1, x2, x3, x4, y1, y2, y3, y4)
}
