% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robertsons_chart.R
\name{QtnA}
\alias{QtnA}
\title{Fr-Qtn relationship for boundary zone 1}
\usage{
QtnA(Fr)
}
\arguments{
\item{Fr}{Friction ratio, in percentage}
}
\value{
Qtn values
}
\description{
Calculate Qtn based on known Fr for line seperating zone 1 from the
other zones
}
