% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robertsons_chart.R
\name{QtnG}
\alias{QtnG}
\title{Fr-Qtn relationship for boundary zone 8 and 9}
\usage{
QtnG(Fr)
}
\arguments{
\item{Fr}{Friction ratio, in percentage}
}
\value{
Qtn values
}
\description{
Calculate Qtn based on known Fr for line seperating zone 8 and 9 from the
other zones
}
