% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robertsons_chart.R
\name{robertson_assign_zone}
\alias{robertson_assign_zone}
\title{Assign Robertson's soil type based in Fr and Qtn values}
\usage{
robertson_assign_zone(Fr, Qtn)
}
\arguments{
\item{Fr}{Friction ratio, in \%}

\item{Qtn}{Normalised tip resistance}
}
\value{
array with soil type numbers
}
\description{
Assign Robertson's soil type based in Fr and Qtn values. Function returns
the number of the soil type, i.e. 1 - 9
}
