% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robertsons_chart.R
\name{robertson_polygons}
\alias{robertson_polygons}
\title{Create polygons for each soil zone in Robertson's chart}
\usage{
robertson_polygons(nFr = 21, Fr_lim = c(0.1, 10), Qtn_lim = c(1, 1000))
}
\arguments{
\item{nFr}{number of points on curved polygon edges. Fr interval is split
linearly}

\item{Fr_lim}{Fr lower and upper limit of plot}

\item{Qtn_lim}{Qtn lower and upper limit of plot}
}
\value{
tibble with Fr (field `Fr`), Qtn (field `Qtn`), zone numbers
  (field `zone`) and soil descriptions (field `soil`)
}
\description{
Generate polygons defined by (Frn, Qtn) points for each soil zone in
Robertson's chart
}
