% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohr_circles.R
\name{rotate_elements}
\alias{rotate_elements}
\title{Function to rotate coordinates}
\usage{
rotate_elements(x, y, theta = 0)
}
\arguments{
\item{x}{x-coordinate (array)}

\item{y}{y-coordinate (array)}

\item{theta}{rotation angle (array, in radians)}
}
\value{
a tibble with rotated positions in fields `x2` and `y2`
}
\description{
Simple rotation function used for plotting rotated soil element
and stress arrows
}
