% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohr_circles.R
\name{rotate_stresses}
\alias{rotate_stresses}
\title{Function to calculate rotated stresses}
\usage{
rotate_stresses(sigx, sigz, tau, theta)
}
\arguments{
\item{sigx}{normal stress on x-plane (array)}

\item{sigz}{normal stress on z-plane (array)}

\item{tau}{shear stress (array)}

\item{theta}{rotation angle (array, in radians)}
}
\value{
a tibble with rotated stresses in fields `sigx2`, `sigy2` and `tau2`
}
\description{
Function to apply stress rotation
}
