% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plotting.R
\name{round_limits}
\alias{round_limits}
\title{Round values in an array to a 'nice' nearby numbers in plots}
\usage{
round_limits(
  vals,
  lower = NA,
  upper = NA,
  log = FALSE,
  vals_round = c(1, 2, 5, 10)
)
}
\arguments{
\item{vals}{values to be bounded in array}

\item{lower}{user-defined lower limit, which overrides limit found}

\item{upper}{user-defined upper limit, which overrides limit found}

\item{log}{if `log == FALSE`, a linear scale is assumed. If `log == TRUE`,
a log10 scale is assumed}

\item{vals_round}{an array with potential rounding values, after values
have been scaled by their order of magnitude. Optional values have to be
on the domain `1 <= vals_round < 10`}
}
\value{
two-value array with lower and upper limit
}
\description{
Function to round an array of values to a 'nice' nearby numbers rounded
numbers. This can be useful for setting plot limits. Function returns
a bounding range
}
\examples{
x <- c(51, 321)
round_limits(x)
round_limits(x, lower = 0)
round_limits(x, log = TRUE)
}
