% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R
\name{solve_flownet}
\alias{solve_flownet}
\title{Solve flow net head problem for concatenated rectangular domains}
\usage{
solve_flownet(df)
}
\arguments{
\item{df}{tibble with problem description, generated by function
`generate_flownet_properties()`}
}
\value{
a tibble with heads (field `h`) and flow rates in x and y-direction
  (`qx`, `qy`) for each real point in the finite difference grid. Each point
  is characterised by a position (`x`, `y`) and an index `id` indicating which
  domain the point belongs to
}
\description{
Solves a flow net problem for a problem consisting of connected rectangular
grids. This function solves the hydroloc head using boundary conditions and
a finite difference approach.
}
\examples{
df <- generate_flownet_properties()
dp <- solve_flownet(df)
}
