% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_rectangular.R
\name{solve_flownet_rectangular}
\alias{solve_flownet_rectangular}
\title{Solve flow net head problem for concatenated rectangular domains}
\usage{
solve_flownet_rectangular(df)
}
\arguments{
\item{df}{tibble with problem description, generated by function
`flownet_geometry_rectangular()`}
}
\value{
a tibble with heads (field `h`), flow rates in x and y-direction
  (`qx`, `qy`) and flow potential (`psi`) for each real point in the finite
  difference grid. Each point is characterised by a position (`x`, `y`)
  and an index `id` indicating which domain the point belongs to
}
\description{
Solves a flow net problem for a problem consisting of connected rectangular
grids. This function solves the hydraulic head using boundary conditions and
a finite difference approach, and also calculates the flow potential psi
}
\examples{
df <- flownet_geometry_rectangular()
dp <- solve_flownet_rectangular(df)
}
