% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_profiles_vertical.R
\name{tabulate_vertical_stress_profile}
\alias{tabulate_vertical_stress_profile}
\title{Generate bookdown table for stress calculatinos}
\usage{
tabulate_vertical_stress_profile(
  ds,
  output = "values",
  traces = c("sigma_v", "u", "sigma'_v"),
  nround = 2
)
}
\arguments{
\item{ds}{tibble with stress values. Output from function
`vertical_stress_profile()`}

\item{output}{if `values`, only return stress values. If `calculations`,
it also return how to calculate the stresses}

\item{traces}{array with traces to be shown. May contain `sigma_v` for
the total vertical stress, `u` for the pore water pressure and `sigma'_v`
for the vertical effective stress}

\item{nround}{number of decimals to round stress values to}
}
\value{
a tibble
}
\description{
Generates a tibble that can be converted using knitr to a nicely looking
bookdown table
}
\examples{
ds <- vertical_stress_profile(z_watertable = -2, q = 10)
tabulate_vertical_stress_profile(ds)
}
