% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_profiles_vertical.R
\name{vertical_stress_profile}
\alias{vertical_stress_profile}
\title{Calculate verticals stresses using layers}
\usage{
vertical_stress_profile(
  z_soil = c(0, 2, 4),
  z_watertable = -2,
  z_max = 6,
  gamma_b = c(17.5, 15, 20),
  gamma_w = 10,
  q = 10,
  ...
)
}
\arguments{
\item{z_soil}{depth of top of each layer}

\item{z_watertable}{depth of the water table}

\item{z_max}{maximum depth of interest}

\item{gamma_b}{unit weight of soil, in kN/m3. Should be a scalar or a
vector with the same length as `z_soil`}

\item{gamma_w}{unit weight of water: 10 kN/m3}

\item{q}{surcharge at soil surface}

\item{...}{additional fields to be added to the returned data}

\item{z_interval}{additional interval at which to calculate stresses}
}
\value{
a tibble with depth `z`, vertical total stress `sigma_v`, vertical
  effective stress `sigma'_v` and pore pressure `u`. All fields in `...`
  are added too
}
\description{
Function takes a soil profile and returns the vertical stresses at all
layer interfaces
}
\examples{
#only calculate stresses at layer interfaces and water table
vertical_stress_profile(
  c(0, 4, 6),
  z_watertable = 2,
  gamma_b = c(18, 20, 22),
  q = 10
)

#add stresses at regular intervals
vertical_stress_profile(
  c(0, 4, 6),
  z_watertable = 2,
  gamma_b = c(18, 20, 22),
  q = 10,
  z_interval = 1
)
}
