% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_plotting.R
\name{amount_equipotentialintervals}
\alias{amount_equipotentialintervals}
\title{Decide on number of equipotential intervals}
\usage{
amount_equipotentialintervals(df, dp, Nf = 5)
}
\arguments{
\item{df}{tibble with information about domain}

\item{dp}{tibble with head and flow potential results for each node}

\item{Nf}{number of requested flow channels}
}
\value{
the optimal number of equipotential intervals `Nd`
}
\description{
This function takes a finite difference solution for both head `h` and flow
potential function `psi` to get the best number of equipotential intervals
`Nd` based on a given number of flow channels `Nf`. This function assumes
the permeabilities in all domains is the same
}
\examples{
df <- flownet_geometry_rectangular()
dp <- flownet_solve_rectangular(df)
amount_equipotentialintervals(df, dp, Nf = 5)
}
