% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_net.R
\name{generate_flownet_properties}
\alias{generate_flownet_properties}
\title{Generate tibble describing flow net problem}
\usage{
generate_flownet_properties(
  x0 = 0,
  y0 = 0,
  Lx = c(15, 15, 5, 20, 20),
  Ly = c(7.5, 10, 10, 10, 5),
  Ly_dry = c(0, 0, 0, 0, 0),
  kx = 1e-06,
  ky = 1e-06,
  type1 = c("q", "q", "c", "c", "q"),
  type2 = c("h", "c", "q", "c", "h"),
  type3 = c("q", "c", "c", "q", "q"),
  type4 = c("c", "q", "q", "q", "c"),
  value1 = c(0, 0, 2, 3, 0),
  value2 = c(20, 1, 0, 5, 18),
  value3 = c(0, 3, 4, 0, 0),
  value4 = c(2, 0, 0, 0, 4),
  grid_size = 0.2,
  nodes_min = 3
)
}
\arguments{
\item{x0, y0}{x and y position of the bottom left corner of the first domain}

\item{Lx, Ly}{size in x and y direction of each rectangular domain}

\item{Ly_dry}{thickness of any dry soil directly on top of this domain.
This is only used to generate the soil geometry later. This extra height
is only plotted when the top edge of the domain (edge 2) has a fixed
head as boundary condition}

\item{kx, ky}{permeabilities in x and y-direction. If scalars, they are set
equal for each domain}

\item{type1, type2, type3, type4}{Type of boundary condition on each of the
four edges of each domain. Edges are numbered clockwise:
\itemize{
\item \code{type1} for the left edge (negative x-face)
\item \code{type2} for the top edge (positive y-face)
\item \code{type3} for the right edge (positive x-face)
\item \code{type4} for the bottom edge (negative y-face)
Possible values for the boundary condition types are:
\item \code{type = "q"}: known flow (positive in inflow)
\item \code{type = "h"}: known head
\item \code{type = "c"}: face is connected to a face in another domain
}}

\item{value1, value2, value3, value4}{values of boundary conditions:
\itemize{
\item if \code{type = "q"}, \code{value} is the flow rate in positive direction
\item if \code{type = "h"}: \code{value} is the known head
\item if \code{type = "c"}: \code{value} gives the number of the domain that the face
is connected to (domains are numbered 1, 2, 3, in order of appearance)
}}

\item{grid_size}{requested finite difference grid size. The grid size used
may vary slighly per domain, in order to fit a integer number of nodes in
the domain with a given size}

\item{nodes_min}{the minimum number of real nodes in any direction in any
domain. Set to at least 3 to allow all finite difference approximations
to work}
}
\value{
a tibble describing the entire tibble. In addition to the input,
fields are added for \code{x0} and \code{y0} (the x,y position of the bottom left)
point in each domain, \code{nx} and \code{ny} (the number of real nodes in each
domain in each direction), and \code{hx}, and \code{hy} (the distance between
nodes in each direction)
}
\description{
The flow net problem consists of connected rectangular domains. The
permabilities can be set seperately for each domain.
}
\examples{
#create the default flow net
generate_flownet_properties()
}
