% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plotting.R
\name{ggplot_add_soilmarker}
\alias{ggplot_add_soilmarker}
\title{Annotate soil surface marker to a ggplot object}
\usage{
ggplot_add_soilmarker(
  plt,
  xc,
  yc,
  scale = 1,
  line_size = 0.5,
  colour_soil = "#65571d",
  n = 3
)
}
\arguments{
\item{plt}{ggplot object}

\item{xc, yc}{x and y-position where marker touches the soil surface (middle)}

\item{scale}{scaling factor for marker. Default width is approx 1}

\item{line_size}{thickness of lines}

\item{colour_soil}{line colour for all lines}

\item{n}{number of lines to plot underneath the water table, as part of the symbol}
}
\value{
ggplot object with added water table marker
}
\description{
Adds a standard soil surface indicator icon to an existing ggplot object
}
\examples{
#generate simple plot
plt <- ggplot2::ggplot() +
  ggplot2::annotate("rect", xmin = -2, xmax = 2, ymin = -2, ymax = 2, fill = "yellow") +
  ggplot2::coord_fixed(ratio = 1)

#add water table marker
ggplot_add_soilmarker(plt, x = 1, y = 1, scale = 1)
}
