% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_plotting.R
\name{ggplot_add_watermarker}
\alias{ggplot_add_watermarker}
\title{Annotate water table marker to a ggplot object}
\usage{
ggplot_add_watermarker(
  plt,
  xc,
  yc,
  scale = 1,
  line_size = 0.5,
  fill_water = "#2a7fff",
  colour_water = "#000080"
)
}
\arguments{
\item{plt}{ggplot object}

\item{scale}{scaling factor for marker. Default height is approx 1}

\item{line_size}{thickness of lines}

\item{fill_water}{fill colour for triangle}

\item{colour_water}{line colour for all lines}

\item{x, y}{x and y-position where marker touches water table}
}
\value{
ggplot object with added water table marker
}
\description{
Adds a standard water table marker (upside-down triangle over a number of
horizontal lines) to an existing ggplot
}
\examples{
#generate simple plot
plt <- ggplot2::ggplot() +
  ggplot2::annotate("rect", xmin = -2, xmax = 2, ymin = -2, ymax = 2, fill = "yellow") +
  ggplot2::coord_fixed(ratio = 1)

#add water table marker
ggplot_add_watermarker(plt, x = 1, y = 1, scale = 1)
}
