% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_net.R
\name{ggplot_flownet_dam}
\alias{ggplot_flownet_dam}
\title{Plot flow net and soil geometry for a dam/sheet pile problem}
\usage{
ggplot_flownet_dam(
  df,
  dp = NULL,
  Nf = 5,
  xlab = "x [m]",
  ylab = "z [m]",
  fill_soilwet = "#aebab7",
  fill_soildry = "#d3bc5f",
  colour_soil = "#65571d",
  colour_flowline = "black",
  colour_equipotentialline = "black",
  fill_water = "#2a7fff",
  colour_water = "#000080",
  fill_impermeable = "#333333",
  fill_air = "white",
  line_water = 2,
  line_width = 0.5,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  axes = TRUE,
  thickness_impermeable = c(0, 1, 0, 1),
  scale_soilmarker = 0.075,
  scale_watermarker = 0.05
)
}
\arguments{
\item{df}{tibble with domains/problem description, generated by function
`generate_flownet_properties()`.#'}

\item{dp}{solution for heads and flow potential for each node. If not
defined, it is calculated in this plotting function}

\item{Nf}{number of requested flow paths}

\item{xlab, ylab}{x and y-axis label}

\item{fill_soilwet}{fill colour for saturated soil}

\item{fill_soildry}{fill colour for dry soil,}

\item{colour_soil}{line colour for soil}

\item{colour_flowline}{colour of flow lines}

\item{colour_equipotentialline}{colour of equipotential lines}

\item{fill_water}{fill colour for water}

\item{colour_water}{line colour for water}

\item{fill_impermeable}{fill colour for impermeable parts of the domain}

\item{fill_air}{fill colour for air on top of soil/water}

\item{line_water}{line type for water edges (defined head)}

\item{line_width}{line thickness for all lines}

\item{xlim, ylim}{user-defined axis limits}

\item{axes}{if `FALSE` no axes are plotted}

\item{thickness_impermeable}{thickness of impermeable border around the
entire problem. Defined as an array for all edges (left, top, right,
bottom), or single value for all edges}

\item{scale_soilmarker}{size of soil marker, in fraction of height}

\item{scale_watermarker}{size of water table marker, in fraction of height}
}
\value{
a ggplot object
}
\description{
Plot flow net and soil geometry for a dam/sheet pile problem
}
\examples{
#default example
df <- generate_flownet_properties()
ggplot_flownet_dam(df)

#anisotropic permeability - 6 flow channels
df <- generate_flownet_properties(kx = 4e-6, ky = 1e-6)
ggplot_flownet_dam(df, Nf = 6)

#excavation next to river
df <- generate_flownet_properties(
  Lx = c(20, 20, 0.2, 5, 5),
  Ly = c(10, 5, 5, 5, 2),
  Ly_dry = c(2, 0, 0, 0, 0),
  type1 = c("q", "q", "c", "c", "q"),
  type2 = c("h", "c", "q", "c", "h"),
  type3 = c("q", "c", "c", "q", "q"),
  type4 = c("c", "q", "q", "q", "c"),
  value1 = c(0, 0, 2, 3, 0),
  value2 = c(12, 1, 0, 5, 6),
  value3 = c(0, 3, 4, 0, 0),
  value4 = c(2, 0, 0, 0, 4)
)
ggplot_flownet_dam(df)
}
