% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_differences.R
\name{index_grid2vector}
\alias{index_grid2vector}
\title{Convert node grid positions to indices}
\usage{
index_grid2vector(ix, iy, nx, ny, i0 = 0, real_only = FALSE, ...)
}
\arguments{
\item{ix, iy}{arrays with index of position in x and y directions}

\item{nx}{number of real nodes in x and y directions}

\item{ny}{number of real nodes in x and y directions}

\item{i0}{node index offset}

\item{real_only}{if `TRUE`, only real nodes are numbered and ghost nodes
are ignored in numbering}

\item{...}{additional arguments to pass}
}
\value{
array with node index for each node
}
\description{
Function takes a vectors of x and y node indices and returns the index
(numbered row-wise)
}
\examples{
#define grid
nx <- 4
ny <- 3

#including ghost nodes
i <- seq(nodes_total(nx, ny))
ixy <- index_vector2grid(i, nx, ny)
index_grid2vector(ixy$ix, ixy$iy, nx, ny)

#excluding ghost nodes
i <- seq(nodes_total(nx, ny, real_only = TRUE))
ixy <- index_vector2grid(i, nx, ny, real_only = TRUE)
index_grid2vector(ixy$ix, ixy$iy, nx, ny, real_only = TRUE)
}
