% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R
\name{index_real}
\alias{index_real}
\title{Get indices of all real nodes}
\usage{
index_real(nx, ny, i0 = 0, ...)
}
\arguments{
\item{nx}{number of real nodes in x and y-direction}

\item{ny}{number of real nodes in x and y-direction}

\item{i0}{node starting number.}

\item{...}{potential extra arguments. Required to use the fuction with the
`purrr::pmap()` function}
}
\value{
array of indices for all real nodes (non-ghost nodes)
}
\description{
Get an array of the indices of all real nodes in a rectangular grid of
(real) nodes. Nodes are number per row --> per column. Ghost nodes
are placed outside the edges of all sides of the rectangular grid
}
