% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flownet_finitedifferences.R
\name{index_vector2grid}
\alias{index_vector2grid}
\title{Convert node position to grid x-y position}
\usage{
index_vector2grid(i, nx, ny, i0 = 0, real_only = FALSE, ...)
}
\arguments{
\item{i}{array with node indices}

\item{nx, ny}{number of real nodes in x and y directions}

\item{i0}{node index offset}

\item{real_only}{if `TRUE`, only real nodes are numbered and ghost nodes
are ignored in numbering}

\item{...}{additional arguments to pass}
}
\value{
a list with two fields: `ix` and `iy` with indices in x and y
  directions
}
\description{
Function takes a vector of node indices (numbered row-wise) and
returns a list of their x and y indices.
}
\examples{
#convert to grid positions
nx <- 4
ny <- 3
i <- seq(1, (nx + 2)*(ny + 2) - 4)
index_vector2grid(i, nx, ny)
}
