% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R
\name{nodal_coordinates_real}
\alias{nodal_coordinates_real}
\title{Get coordinates of all real nodes in the finite difference grid}
\usage{
nodal_coordinates_real(df)
}
\arguments{
\item{df}{tibble with all properties for the rectangular domains}
}
\value{
a tibble with fields `x` and `y` for positions, and `id` to indicate
  which grid the point belongs to
}
\description{
Function obtains the x,y positions of all real nodes in the finite difference
grid
}
