% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_differences.R, R/flow_nets.R
\name{nodal_coordinates_real}
\alias{nodal_coordinates_real}
\title{Get coordinates of all real nodes in the finite difference grid}
\usage{
nodal_coordinates_real(df)

nodal_coordinates_real(df)
}
\arguments{
\item{df}{tibble with all properties for the rectangular domains}

\item{nx, ny}{number of real nodes on the grid}

\item{Lx, Ly}{length of grid in x and y direction}

\item{x0, y0}{position of origin of the grid}
}
\value{
a tibble with fields `x` and `y` for positions, and `id` to indicate
  which grid the point belongs to

a tibble with fields `x` and `y` for positions, and `id` to indicate
  which grid the point belongs to
}
\description{
Function obtains the x,y positions of all real nodes in the finite difference
grid

Function obtains the x,y positions of all real nodes in the finite difference
grid
}
\examples{
df <- nodal_coordinates_real(
  nx = c(4, 5),
  ny = c(3, 4),
  Lx = c(3, 2),
  Ly = c(2, 3),
  x0 = c(0, 4),
  y0 = c(0, 4)
)
plot(df$x, df$y, "b")
}
