% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_nets.R
\name{nodes_total}
\alias{nodes_total}
\title{Determine total number of nodes}
\usage{
nodes_total(nx, ny, real = FALSE)
}
\arguments{
\item{nx}{number of real nodes in x and y-direction}

\item{ny}{number of real nodes in x and y-direction}

\item{if}{`real = FALSE`, all nodes, including ghost nodes are counted.
if `real = TRUE`, only real nodes are counted}
}
\value{
number of nodes (scalar)
}
\description{
Determine the total number of nodes on a grid of nodes, including any
ghost nodes outside all four sides of the grid
}
