% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_differences.R
\name{nodes_total}
\alias{nodes_total}
\title{Determine total number of nodes}
\usage{
nodes_total(nx, ny, real_only = FALSE, ...)
}
\arguments{
\item{nx}{number of real nodes in x and y directions}

\item{ny}{number of real nodes in x and y directions}

\item{real_only}{if `TRUE`, only real nodes are numbered and ghost nodes
are ignored in numbering}

\item{...}{additional arguments to pass}
}
\value{
number of nodes (scalar)
}
\description{
Determine the total number of nodes on a grid of nodes. Ghost nodes
are or are not included based on `real_only` setting.
}
